% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{apply}
\alias{apply}
\title{Apply Functions Over matrix Margins}
\usage{
apply(X, MARGIN, FUN, ..., simplify = TRUE)
}
\arguments{
\item{X}{An \code{IterableMatrix} object.}

\item{MARGIN}{A single number giving the subscripts which the function will
be applied over, \code{1} indicates rows, \code{2} indicates columns.}

\item{FUN}{Function to be applied. \code{FUN} is found by a call to \link{match.fun}
and typically is either a function or a symbol (e.g., a backquoted name) or a
character string specifying a function to be searched for from the
environment of the call to apply.}

\item{...}{optional arguments to \code{FUN}.}

\item{simplify}{a logical indicating whether results should be
    simplified if possible.}
}
\value{
If each call to \code{FUN} returns a vector of length \code{n}, and simplify is \code{TRUE},
then \code{apply} returns an array of dimension \code{c(n, dim(X)[MARGIN])} if \code{n > 1}.
If \code{n} equals \code{1}, \code{apply} returns a vector if \code{MARGIN} has length 1 and an
array of dimension \code{dim(X)[MARGIN]} otherwise. If \code{n} is \code{0}, the result has
length 0 but not necessarily the ‘correct’ dimension.

If the calls to \code{FUN} return vectors of different lengths, or if \code{simplify}
is \code{FALSE}, \code{apply} returns a list of length \code{dim(X)[MARGIN]}.
}
\description{
Apply Functions Over matrix Margins
}
