% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_project_predict.R
\name{c(PipelineBase)}
\alias{c(PipelineBase)}
\title{Combine pipeline objects, to create a new pipeline object.}
\usage{
\S4method{c}{PipelineBase}(x, ...)
}
\arguments{
\item{x}{(PipelineBase) The pipeline object to combine to.}

\item{...}{(PipelineBase) The pipeline objects to combine from.}
}
\description{
Combine pipeline objects, to create a new pipeline object.
}
