% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_pipelines.R
\name{check_pipeline_validity}
\alias{check_pipeline_validity}
\title{Check if all steps are transformers, with the final step being either an estimator or a transformer}
\usage{
check_pipeline_validity(object)
}
\arguments{
\item{object}{(Pipeline) The pipeline object to check}
}
\description{
Check if all steps are transformers, with the final step being either an estimator or a transformer
}
\keyword{internal}
