% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{checksum}
\alias{checksum}
\title{Calculate the MD5 checksum of an IterableMatrix}
\usage{
checksum(matrix)
}
\arguments{
\item{matrix}{IterableMatrix object}
}
\value{
MD5 checksum string in hexidecimal format.
}
\description{
Calculate the MD5 checksum of an IterableMatrix and return the checksum in
hexidecimal format.
}
\details{
\code{checksum()} converts the non-zero elements of the sparse input matrix to double
precision, concatenates each element value with the element row and column index words,
and uses these 16-byte blocks along with the matrix dimensions and row and column
names to calculate the checksum. The checksum value depends on the storage order so
column- and row-order matrices with the same element values give different checksum
values. \code{checksum()} uses element and index values in little-endian CPU storage order.
It converts to little-endian order on big-endian architecture although this has not
been tested.
}
\examples{
library(Matrix)
library(BPCells)
m1 <- matrix(seq(1,12), nrow=3)
m2 <- as(m1, 'dgCMatrix')
m3 <- as(m2, 'IterableMatrix')
checksum(m3)
}
