% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_membership_matrix}
\alias{cluster_membership_matrix}
\title{Convert grouping vector to sparse matrix}
\usage{
cluster_membership_matrix(groups, group_order = NULL)
}
\arguments{
\item{groups}{Vector with one entry per cell, specifying the cell's group}

\item{group_order}{Optional vector listing ordering of groups}
}
\value{
cell x group matrix where an entry is 1 when a cell is in a given group
}
\description{
Converts a vector of membership IDs into a sparse matrix
}
