% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_stats.R
\name{colQuantiles}
\alias{colQuantiles}
\title{Find the nth quantile value(s) of each column in a matrix. Only supports non-transposed matrices.}
\usage{
colQuantiles(
  x,
  rows = NULL,
  cols = NULL,
  probs = seq(from = 0, to = 1, by = 0.25),
  na.rm = FALSE,
  type = 7L,
  ...,
  useNames = TRUE,
  drop = TRUE
)
}
\arguments{
\item{x}{IterableMatrix object or a matrix-like object.}

\item{probs}{(Numeric) Quantile value(s) to be computed, between 0 and 1.}

\item{type}{(Integer) between 4 and 9 selecting which quantile algorithm to use, detailed in \code{matrixStats::rowQuantiles()}}
}
\value{
\itemize{
\item \verb{colQuantiles():} If \code{length(probs) == 1}, return a numeric with number of entries equal to the number of columns in the matrix.
Else, return a Matrix of quantile values, with cols representing each quantile, and each row representing a col in the input matrix.
}
}
\description{
Find the nth quantile value(s) of each column in a matrix. Only supports non-transposed matrices.
}
