% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{concat_dimnames}
\alias{concat_dimnames}
\title{Helper function for rbind/cbind concatenating dimnames}
\usage{
concat_dimnames(x, y, len_x, len_y, warning_prefix, dim_type)
}
\description{
Helper function for rbind/cbind concatenating dimnames
}
\keyword{internal}
