% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{convert_mat_to_cluster_matrix}
\alias{convert_mat_to_cluster_matrix}
\title{Convert a matrix to a required input type for a clustering step}
\usage{
convert_mat_to_cluster_matrix(
  mat,
  required_mat_type = c("knn", "adjacency"),
  knn_mat_method = knn_annoy,
  knn_graph_method = knn_to_geodesic_graph
)
}
\arguments{
\item{mat}{Input matrix to be converted}

\item{required_mat_type}{(character) Type of matrix required for clustering. Must be one of "adjacency" or "knn".}

\item{knn_mat_method}{(function) Function to convert input matrix to knn matrix. Must be a (optionally partialized)
version of \code{knn_hnsw()} or \code{knn_annoy()}.  Unused if mat is already a knn or graph matrix.}

\item{knn_graph_method}{(function) Function to convert input matrix to knn graph. Must be a (optionally partialized)
version of \code{knn_to_graph()}, \code{knn_to_snn_graph()} or \code{knn_to_geodesic_graph()}.  Unused if required_mat_type is "knn",
or if mat is already a graph matrix.}
}
\description{
Ensure that if a user is using a clustering function, the input matrix is converted to a knn or graph matrix.
Returns the input matrix if it is already the correct type, otherwise converts it to the correct type.
If the required matrix is an adjacency matrix, the input matrix will optionally be converted to a knn matrix if it is not one already.
}
\details{
Clustering steps in BPCells use the following pattern:
\enumerate{
\item The input matrix is converted to a knn matrix using a function named \verb{knn_\{hnsw, annoy\}()}
\item The knn matrix is converted to a graph matrix using one of the functions named \verb{knn_to_*_graph()}
\item The graph matrix is clustered using a function named \verb{cluster_graph_*()}.
}

Each one of these steps requires a specific type of matrix as input.
The typing of the input matrices is determined by the \code{mat_type} attribute given by the knn_to_*_graph() and \verb{knn_\{hnsw, annoy\}()}
functions,as returned data types vary between all clustering functions.
}
\keyword{internal}
