% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{convert_mat_to_graph}
\alias{convert_mat_to_graph}
\title{Converts a matrix to a graph adjacency matrix}
\usage{
convert_mat_to_graph(
  mat,
  knn_method = knn_hnsw,
  knn_to_graph_method = knn_to_geodesic_graph,
  threads = 1L,
  verbose = FALSE
)
}
\arguments{
\item{mat}{Input matrix to be converted of shape \verb{(cells x n_embeddings)}}

\item{knn_method}{(function) Function to convert input to a knn object (e.g., \code{knn_hnsw}, \code{knn_annoy}).
Ignored if \code{mat} is already a knn object or graph matrix.}

\item{knn_to_graph_method}{(function) Function to convert a knn object to a graph adjacency matrix
(e.g., \code{knn_to_geodesic_graph}). Ignored if \code{mat} is already a graph matrix.}
}
\value{
Symmetric graph adjacency matrix (dgCMatrix)
}
\description{
Handles inputs so that cell embeddings and knn objects are converted into graph adjacency matrices.
}
\details{
If \code{mat} is a \verb{(cells x n_embeddings)} matrix then \code{knn_method} is used to convert \code{mat} to a knn object.
Knn objects are then converted into a graph adjacency matrix using \code{knn_to_graph_method}. Also does intermediate checks to ensure
that \code{knn_method} and \code{knn_to_graph} properly converted matrix into a correct output.
}
\keyword{internal}
