% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_partial}
\alias{create_partial}
\title{Helper to create partial functions}
\usage{
create_partial(missing_args = list())
}
\arguments{
\item{missing_args}{(named list\link{bool}) Any named index with a TRUE value
will be treated as missing. Designed to be used in the caller with the
\code{base::missing()} function to detect unspecified arguments with default values,
or to manually specifiy other arguments that should not be specialized}
}
\value{
A \code{bpcells_partial} object (a function with some extra attributes)
}
\description{
Automatically creates a partial application of the caller
function including all non-missing arguments.
}
\keyword{internal}
