% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_demo_mat}
\alias{get_demo_mat}
\alias{get_demo_frags}
\alias{remove_demo_data}
\title{Retrieve BPCells demo data}
\usage{
get_demo_mat()

get_demo_frags()

remove_demo_data()
}
\value{
\itemize{
\item \code{get_demo_mat()}: (IterableMatrix) A \verb{(features x cells)} matrix of shape \verb{(1984 x 2724)}.
}

\itemize{
\item \code{get_demo_frags()}: (IterableFragments) A Fragments object with 2724 cells and fragments on chromosomes 4 and 11.
}

\itemize{
\item \code{remove_demo_data()}: \code{NULL}
}
}
\description{
The demo dataset is a subset of the  10x Genomics PBMC 3k dataset, with the matrix and fragments
filtered to cells with at least 1000 reads.  Both the matrix and the fragments are also subset to only genes on chromosomes 4 and 11.
}
\details{
The first time either \code{get_demo_mat()} are ran \code{get_demo_frags()}, the pre-subsetted
demo data is downloaded and stored in the BPCells data directory  (under \code{file.path(tools::R_user_dir("BPcells", which="data"), "demo_data")}).
Subsequent calls to this function will use the previously downloaded matrix/fragments.
The preperation of this matrix can be reproduced by running the internal function \code{prepare_demo_data()} with \code{directory} set to the BPCells data directory.

In the case that demo data is not pre-downloaded and demo data download fails, \code{prepare_demo_data()} will act
as a fallback.

Both the matrix from \code{get_demo_mat()} and the fragments from \code{get_demo_frags()}
may be removed by running \code{remove_demo_data()}.
\itemize{
\item \code{get_demo_mat()}: Retrieve a 1 MB demo \code{IterableMatrix}, representing a subset of the 10X Genomics PBMC 3k dataset.
}

\itemize{
\item \code{get_demo_frags()}: Retrieve a 12.5 MB demo \code{IterableFragments} object, representing a subset of the 10X Genomics PBMC 3k dataset.
}

\itemize{
\item \code{remove_demo_data()}: Remove the demo data from the BPCells data directory.
}
}
