% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell_utils.R
\name{DimReduction}
\alias{DimReduction}
\title{Barebones definition of a DimReduction class.}
\usage{
DimReduction(x, fitted_params = list(), ...)
}
\description{
Represents a latent space output of a matrix after a transformation function, with any required information to reproject other inputs using this object.
Child classes should implement a \code{project} method to allow for the projection of other matrices using
the fitted transformation object.
}
\section{Fields}{

\describe{
\item{\code{cell_embeddings}}{(IterableMatrix, dgCMatrix, matrix) The projected data}

\item{\code{fitted_params}}{(list) A list of parameters used for the transformation of a matrix.}
}}

