% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{discrete_palette}
\alias{discrete_palette}
\title{Discrete color palettes}
\usage{
discrete_palette(name, n = 1)
}
\arguments{
\item{name}{Name of the color palette}

\item{n}{Minimum number of colors needed}
}
\description{
These color palettes are derived from the ArchR color palettes, and provide
large sets of distinguishable colors
}
\details{
If the requested number of colors is too large, a new palette will be constructed
via interpolation from the requested palette
}
