% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneScores.R
\name{distanceToNearestDirectional}
\alias{distanceToNearestDirectional}
\title{Find signed distance to nearest genomic ranges}
\usage{
distanceToNearestDirectional(ranges, addArchRBug = FALSE)
}
\arguments{
\item{ranges}{GenomicRanges object}

\item{addArchRBug}{boolean to reproduce ArchR's bug that incorrectly handles nested genes}
}
\value{
A 2-column data.frame with columns upstream and downstream, containing
the distances to the nearest neighbor in the respective directions.
For ranges on + or * strand, distance is calculated as:
\itemize{
\item upstream = max(start(range) - end(upstreamNeighbor), 0)
\item downstream = max(start(downstreamNeighbor) - end(range), 0)
}

For ranges on - strand, the definition of upstream and downstream is flipped.
Note that this definition of distance is slightly different from
\code{GRanges::distance}, as it effectively assumes end-coordinates are not included
in a range.
}
\description{
Given a set of genomic ranges, find the distance to the nearest neighbors both
upstream and downstream.
}
