% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplots.R
\name{draw_trackplot_grid}
\alias{draw_trackplot_grid}
\title{Combine ggplot track plots into an aligned grid.}
\usage{
draw_trackplot_grid(
  ...,
  labels,
  title = NULL,
  heights = rep(1, length(plots)),
  label_width = 0.2,
  label_style = list(fontface = "bold", size = 4)
)
}
\arguments{
\item{...}{Plots in order from top to bottom, generally plain ggplots.
To better accomodate many bulk tracks, patchwork objects which contain multiple
tracks are also accepted. In this case, plot labels will be drawn from the
attribute \verb{$patchwork$labels} if present, rather than the \code{labels} argument.}

\item{labels}{Text labels to display for each track}

\item{title}{Text for overarching title of the plot}

\item{heights}{Relative heights for each component plot. It is suggested to use 1 as standard height of a
pseudobulk track.}

\item{label_width}{Fraction of width that should be used for labels relative to the main track area}

\item{label_style}{Arguments to pass to geom_text to adjust label text style}
}
\value{
A plot object with aligned genome plots. Each aligned row has
the text label, y-axis, and plot body. The relative height of each row is given
by heights. A shared title and x-axis are put at the top.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has been renamed to \code{trackplot_combine()}.
}
\examples{
# See ?trackplot_combine for examples
}
\keyword{internal}
