% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeAnnotations.R
\name{ensure_downloaded}
\alias{ensure_downloaded}
\title{Download a file with a custom timeout}
\usage{
ensure_downloaded(path, backup_url, timeout)
}
\arguments{
\item{path}{Output path to write file}

\item{timeout}{timeout in seconds}

\item{url}{to download from}
}
\description{
Download a file with a custom timeout
}
\keyword{internal}
