% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_pipelines.R
\name{estimate(PipelineBase,IterableMatrix)}
\alias{estimate(PipelineBase,IterableMatrix)}
\title{Estimate predictions on the output data using a fitted pipeline}
\usage{
estimate(object, x, ...)
}
\arguments{
\item{object}{(PipelineBase) The fitted pipeline object. Either the final step is an Estimator, or the pipeline is a single Estimator.}

\item{x}{(IterableMatrix) Input data to be estimated on}
}
\value{
Predicted output labels
}
\description{
Estimate predictions on the output data using a fitted pipeline
}
