% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_pipelines.R
\docType{class}
\name{Estimator}
\alias{Estimator}
\title{PipelineStep representing an operation that estimates data, and holds fitted parameters.}
\description{
PipelineStep representing an operation that estimates data, and holds fitted parameters.
}
\details{
Estimators represent single operations (derived from the PipelineStep class) that make predictions based on data given by an
IterableMatrix.  Additionally, supervised estimators will require a target numeric, or character array to be provided during a \code{fit()} call.
Unsupervised estimators, on the other hand, do not require a target array.  Following a \code{fit()} call, the estimator will hold the fitted parameters
such that data can be labeled using the \code{estimate()} method.

Estimators can be combined into a Pipeline object using the \code{c()} function, with other transformers.  Estimators are required to be the terminal step
within a pipeline.  Estimators can also be combined with full pipelines, to create a new pipeline object.
Derived classes should implement the \code{fit()}, \code{estimate()}, and \code{short_description()} methods.
}
\section{Slots}{

\describe{
\item{\code{step_name}}{(character) Name of the step}

\item{\code{fitted}}{(logical) Whether the pipeline has been fitted}
}}

