% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{EXPERIMENTAL_write_matrix_dir}
\alias{EXPERIMENTAL_write_matrix_dir}
\title{Write to experimental sparse-column format integer matrix}
\usage{
EXPERIMENTAL_write_matrix_dir(mat, dir, buffer_size = 8192L, overwrite = FALSE)
}
\arguments{
\item{dir}{Directory to save the data into}

\item{overwrite}{If \code{TRUE}, write to a temp dir then overwrite existing data. Alternatively,
pass a temp path as a string to customize the temp dir location.}
}
\description{
The experimental sparse-column format is designed to handle storage of matrices
with many columns of all-zero, and less than 2^32-1 non-zero entries.
}
\keyword{internal}
