% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneScores.R
\name{extend_ranges}
\alias{extend_ranges}
\title{Extend genome ranges in a strand-aware fashion.}
\usage{
extend_ranges(
  ranges,
  upstream = 0,
  downstream = 0,
  metadata_cols = c("strand"),
  chromosome_sizes = NULL,
  zero_based_coords = !is(ranges, "GRanges")
)
}
\arguments{
\item{ranges}{Genomic regions given as GRanges, data.frame, or list. See \code{help("genomic-ranges-like")} for details on format and coordinate systems. Required attributes:
\itemize{
\item \code{chr}, \code{start}, \code{end}: genomic position
}}

\item{upstream}{Number of bases to extend each range upstream (negative to shrink width)}

\item{downstream}{Number of bases to extend each range downstream (negative to shrink width)}

\item{metadata_cols}{Optional list of metadata columns to require & extract}

\item{chromosome_sizes}{(optional) Size of chromosomes as a \link{genomic-ranges} object}

\item{zero_based_coords}{If true, coordinates start and 0 and the end coordinate is not included in the range.
If false, coordinates start at 1 and the end coordinate is included in the range}
}
\description{
Extend genome ranges in a strand-aware fashion.
}
\details{
Note that ranges will be blocked from extending past the beginning of the chromosome (base 0),
and if \code{chromosome_sizes} is given then they will also be blocked from extending past the end of the chromosome
}
