% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneScores.R
\name{extendGenomicRanges}
\alias{extendGenomicRanges}
\title{Extend GenomicRanges in a strand-aware fashion.}
\usage{
extendGenomicRanges(ranges, upstream = 0, downstream = 0)
}
\arguments{
\item{ranges}{GenomicRanges object}

\item{upstream}{Number of bases to extend each range upstream (negative to shrink width)}

\item{downstream}{Number of bases to extend each range downstream (negative to shrink width)}
}
\description{
Extend GenomicRanges in a strand-aware fashion.
}
\details{
Note that ranges will be blocked from extending past base 1, or the maximum
}
\keyword{internal}
