% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell_utils.R
\name{select_features_variance}
\alias{select_features_variance}
\alias{select_features_dispersion}
\alias{select_features_mean}
\title{Feature selection functions}
\usage{
select_features_variance(mat, num_feats = 0.05, normalize = NULL, threads = 1L)

select_features_dispersion(
  mat,
  num_feats = 0.05,
  normalize = NULL,
  threads = 1L
)

select_features_mean(mat, num_feats = 0.05, normalize = NULL, threads = 1L)

select_features_by_binned_dispersion(
  mat,
  num_feats = 25000,
  n_bins = 20,
  threads = 1L
)
}
\arguments{
\item{mat}{(IterableMatrix) dimensions features x cells}

\item{num_feats}{(float) Number of features to return.  If the number given is between 0 and 1, treat as a proportion of
the number of rows, rounded down.  Otherwise, treat as an absolute number.
If the number is higher than the number of features in the matrix,
all features will be returned.}

\item{normalize}{(function) Normalize matrix using a given function. If \code{NULL}, no normalization is performed.}

\item{threads}{(integer) Number of threads to use.}

\item{n_bins}{(integer) Number of bins for binning mean gene expression.  Normalizing dispersion is done with respect to each bin,
and if the number of features
within a bin is less than 2, the dispersion is set to 1.}
}
\value{
Return a dataframe with the following columns, sorted descending by score:
\itemize{
\item \code{names}: Feature name.
\item \code{score}: Scoring of the feature, depending on the method used.
\item \code{highly_variable}: Logical vector of whether the feature is highly variable.
}

Each different feature selection method will have a different scoring method:
\itemize{
\item \code{select_features_by_variance}: Score representing variance of each feature.
}

\itemize{
\item \code{select_features_by_dispersion}: Score representing the dispersion of each feature.
}

\itemize{
\item \code{select_features_by_mean}: Score representing the mean accessibility of each feature.
}

\itemize{
\item \code{select_features_by_binned_dispersion}: Score representing the bin normalized dispersion of each feature.
}
}
\description{
Apply a feature selection method to a \verb{(features x cells)} matrix.
}
\details{
\code{select_features_by_variance} Calculates the variance of each feature using the following process:
\enumerate{
\item Perform an optional term frequency + log normalization, for each feature.
\item Find \code{num_feats} features with the highest variance.
}

\code{select_features_by_dispersion} calculates the dispersion of each feature using the following process:
\enumerate{
\item Perform an optional term frequency + log normalization, for each feature.
\item Find the dispersion (variance/mean) of each feature.
\item Find \code{num_feats} features with the highest dispersion.
}

\code{select_features_by_mean} calculates the mean accessibility of each feature using the following process:
\enumerate{
\item Get the sum of each binarized feature.
\item Find \code{num_feats} features with the highest accessibility.
}

\code{select_features_by_binned_dispersion} calculates the bin normalized dispersion of each feature using the following process, given by the Seurat package (Satjia et al. 2015):
\enumerate{
\item Calculate the dispersion of each feature (variance / mean)
\item Log normalize dispersion and mean
\item Bin the features by their means, and normalize dispersion within each bin
}
}
