% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_pipelines.R
\name{fit(PipelineBase,IterableMatrix)}
\alias{fit(PipelineBase,IterableMatrix)}
\title{Fit a pipeline object to data}
\usage{
fit(object, x, y = NULL, ...)
}
\arguments{
\item{object}{(PipelineBase) The pipeline object to fit.}

\item{x}{(IterableMatrix) Input data to be fitted on.}

\item{y}{Optional output data to be fitted on.  Required if the final step is a supervised Estimator, else ignored.}
}
\value{
The fitted pipeline object.
}
\description{
Fit a pipeline object to data
}
\details{
The \code{fit()} method is used to fit a pipeline object to data and a potential label output.  Within single estimators, the \code{fit()} method only
takes the input data to be fitted on. Within pipelines, the \code{fit()} method sequentially fits the transformers on each non-terminal step of the pipeline. More specifically,
The input data is transformed by each transformer, and used to fit the next transformer in the pipeline.  If the final step is an estimator, the input IterableMatrix
and label (if supervised) are used to fit the estimator.

The fitted pipeline object is returned, allowing for projection of new data.
}
