% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac_utils.R
\name{footprint}
\alias{footprint}
\title{Get footprints around a set of genomic coordinates}
\usage{
footprint(
  fragments,
  ranges,
  zero_based_coords = !is(ranges, "GRanges"),
  cell_groups = rlang::rep_along(cellNames(fragments), "all"),
  cell_weights = rlang::rep_along(cell_groups, 1),
  flank = 125L,
  normalization_width = flank\%/\%10L
)
}
\arguments{
\item{fragments}{IterableFragments object}

\item{ranges}{GRanges object with the positions to footprint, or
list/data frame with columns chr, start, & end. For list/data frame, must
include strand information as character vector of "+"/"-", or TRUE/FALSE for
positive/negative strand. "+" strand motifs will footprint around the start
coordinate, and "-" strand motifs will footprint around the end coordinate}

\item{zero_based_coords}{If true, coordinates start and 0 and the end coordinate is not included in the range.
If false, coordinates start at 1 and the end coordinate is included in the range}

\item{cell_groups}{Character or factor assigning a group to each cell, in order of
\code{cellNames(fragments)}}

\item{cell_weights}{Numeric vector assigning weight factors
(e.g. inverse of total reads) to each cell, in order of \code{cellNames(fragments)}}

\item{flank}{Number of flanking basepairs to include on either side of the motif}

\item{normalization_width}{Number of basepairs at the upstream + downstream
extremes to use for normalizing, or 0 to skip normalization}
}
\value{
tibble with columns "group", "position", and "value"
}
\description{
Get footprints around a set of genomic coordinates
}
