% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\docType{methods}
\name{fragment_r_conversion-misc}
\alias{fragment_r_conversion-misc}
\alias{coerce,data.frame,IterableFragments-method}
\alias{coerce,IterableFragments,data.frame-method}
\alias{coerce,GRanges,IterableFragments-method}
\alias{coerce,IterableFragments,GRanges-method}
\title{Convert between BPCells fraagments and R objects misc.}
\usage{
\S4method{coerce}{data.frame,IterableFragments}(from, to, ...)
\S4method{coerce}{IterableFragments,data.frame}(from, to, ...)
\S4method{coerce}{IterableFragments,GRanges}(from, to, ...)
\S4method{coerce}{GRanges,IterableFragments}(from, to, ...)
}
\arguments{
\item{from}{Object supplied to \code{base::coerce()} (typically generated by \code{as()})}

\item{to}{Target class name for coercion}

\item{...}{Additional arguments passed to methods}
}
\description{
See \code{fragment_R_conversion} for main page.
BPCells fragments can be interconverted with GRanges and data.frame R objects.
The main conversion method is R's builtin \code{as()} function, though the convert_to_fragments() helper is also available.
For all R objects except GRanges, BPCells assumes a 0-based, end-exclusive coordinate system.
(See `genomic-ranges-like`` reference for details)
}
\keyword{internal}
