% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{fragment_R_conversion_coercions}
\alias{fragment_R_conversion_coercions}
\alias{as.data.frame.IterableFragments}
\alias{convert_to_fragments}
\title{Convert between BPCells fragments and R objects.}
\usage{
\method{as.data.frame}{IterableFragments}(x, row.names = NULL, optional = FALSE, ...)

convert_to_fragments(x, zero_based_coords = !is(x, "GRanges"))
}
\arguments{
\item{x}{Fragment coordinates given as GRanges, data.frame, or list. See \code{help("genomic-ranges-like")} for details on format and coordinate systems. Required attributes:
\itemize{
\item \code{chr}, \code{start}, \code{end}: genomic position
\item \code{cell_id}: cell barcodes or unique identifiers as string or factor
}}

\item{zero_based_coords}{Whether to convert the ranges from a 1-based end-inclusive
coordinate system to a 0-based end-exclusive coordinate system. Defaults to true
for GRanges and false for other formats
(see this \href{https://web.archive.org/web/20210920203703/http://genome.ucsc.edu/blog/the-ucsc-genome-browser-coordinate-counting-systems/}{archived UCSC blogpost})}

\item{IterableFragments}{BPCells IterableFragments object}

\item{data.frame}{Data frame with columns chr, start, end, and cell_id}

\item{GRanges}{GenomicRanges object with metadata column cell_id}

\item{...}{Additional arguments passed to methods}

\item{row.names}{Optional row names for compatibility with \code{as.data.frame()}.
Ignored for IterableFragments objects.}

\item{optional}{Logical flag for compatibility with \code{as.data.frame()}; ignored.}
}
\value{
\strong{convert_to_fragments()}: IterableFragments object
}
\description{
BPCells fragments can be interconverted with GRanges and data.frame R objects.
The main conversion method is R's builtin \code{as()} function, though the
\code{convert_to_fragments()} helper is also available. For all R objects except
GRanges, BPCells assumes a 0-based, end-exclusive coordinate system. (See
\link{genomic-ranges-like} reference for details)
}
\details{
Coercions rely on base R's \code{as()}; for example \code{as(frags, "data.frame")}
converts BPCells fragments back to a tabular format, while \code{as(x, "IterableFragments")}
materialises supported R objects as fragment stores. Coercions to and from
\code{GRanges} require the GenomicRanges package to be installed.
}
\examples{
frags_table <- tibble::tibble(
  chr = paste0("chr", 1:10),
  start = 0,
  end = 5,
  cell_id = "cell1"
)
frags_table

frags_granges <- GenomicRanges::makeGRangesFromDataFrame(
 frags_table, keep.extra.columns = TRUE
)
frags_granges

#######################################################################
## convert_to_fragments() example
#######################################################################
frags <- convert_to_fragments(frags_granges)
frags


#######################################################################
## as(x, "IterableFragments") example
#######################################################################
frags <- as(frags_table, "IterableFragments")
frags


#######################################################################
## as(bpcells_fragments, "data.frame") example
#######################################################################
frags_table <- as(frags, "data.frame")
frags_table


#######################################################################
## as(bpcells_fragments, "GRanges") example
#######################################################################
frags_granges <- as(frags, "GRanges")
frags_granges


}
