% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{fragments_identical}
\alias{fragments_identical}
\title{Check if two fragments objects are identical}
\usage{
fragments_identical(fragments1, fragments2)
}
\arguments{
\item{fragments1}{First IterableFragments to compare}

\item{fragments2}{Second IterableFragments to compare}
}
\value{
boolean for whether the fragments objects are identical
}
\description{
Check if two fragments objects are identical
}
\examples{
## Prep data
frags_1 <- tibble::tibble(
  chr = "chr1",
  start = seq(10, 260, 50),
  end = start + seq(5, 30, 5),
  cell_id = paste0("cell", c(rep(1, 2), rep(2,2), rep(3,2)))
) \%>\% convert_to_fragments()
frags_1

frags_2_identical <- tibble::tibble(
  chr = "chr1",
  start = seq(10, 260, 50),
  end = start + seq(5, 30, 5),
  cell_id = paste0("cell", c(rep(1, 2), rep(2,2), rep(3,2)))
) \%>\% convert_to_fragments()

frags_3_different <- tibble::tibble(
  chr = "chr1",
  start = seq(10, 260, 50),
  end = start + seq(5, 30, 5),
  cell_id = paste0("cell", c(rep(1, 2), rep(2,2), rep(4,2)))
) \%>\% convert_to_fragments()


## In the case of mismatching cell ids
fragments_identical(frags_1, frags_3_different)


## In the case of two identical frag objects
fragments_identical(frags_1, frags_2_identical)
}
