% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_gene_mapping}
\alias{human_gene_mapping}
\alias{mouse_gene_mapping}
\title{Gene Symbol Mapping data}
\format{
\strong{human_gene_mapping}

A named character vector. Names are aliases or IDs and values
are the corresponding canonical gene symbol

\strong{mouse_gene_mapping}

A named character vector. Names are aliases or IDs and values
are the corresponding canonical gene symbol
}
\source{
\strong{human_gene_mapping}

\url{http://ftp.ebi.ac.uk/pub/databases/genenames/hgnc/tsv/non_alt_loci_set.txt}

\strong{mouse_gene_mapping}

\url{http://www.informatics.jax.org/downloads/reports/MGI_EntrezGene.rpt}
\url{http://www.informatics.jax.org/downloads/reports/MRK_ENSEMBL.rpt}
}
\usage{
human_gene_mapping

mouse_gene_mapping
}
\description{
Mapping of the canonical gene symbols corresponding to each
unambiguous alias, previous symbol, ensembl ID, or entrez ID.
}
\details{
See the source code in \code{data-raw/human_gene_mapping.R} and
\code{data-raw/mouse_gene_mapping.R} for exactly how these mappings were made.
}
\keyword{datasets}
