% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeAnnotations.R
\name{match_gene_symbol}
\alias{match_gene_symbol}
\alias{canonical_gene_symbol}
\title{Gene symbol matching}
\usage{
match_gene_symbol(query, subject, gene_mapping = human_gene_mapping)

canonical_gene_symbol(query, gene_mapping = human_gene_mapping)
}
\arguments{
\item{query}{Character vector of gene symbols or IDs}

\item{subject}{Vector of gene symbols or IDs to index into}

\item{gene_mapping}{Named vector where names are gene symbols or IDs and values are
canonical gene symbols}
}
\value{
\strong{match_gene_symbol}

Integer vector of indices \code{v} such that \code{subject[v]} corresponds to the gene
symbols in \code{query}

\strong{canonical_gene_symbol}

Character vector of canonical gene symbols for each symbol in \code{query}
}
\description{
Correct alias gene symbols, Ensembl IDs, and Entrez IDs to canonical gene
symbols. This is useful for matching gene names between different datasets
which might not always use the same gene naming conventions.
}
