% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneScores2.R
\name{gene_score_tiles}
\alias{gene_score_tiles}
\title{Calculate gene-tile distances for ArchR gene activities}
\usage{
gene_score_tiles(
  genes,
  chromosome_sizes = NULL,
  tile_size = 500,
  addArchRBug = FALSE
)
}
\arguments{
\item{genes}{GRanges object with gene start, end, and strand}

\item{chromosome_sizes}{(optional) Size of chromosomes as a \link{genomic-ranges} object}

\item{tile_size}{Size of tiles to consider}

\item{addArchRBug}{Replicate ArchR bug in handling nested genes}
}
\value{
GRanges object with one range per tile, with additional metadata
columns gene_idx (row index of the gene this tile corresponds to) and
distance.

Distance is a signed distance calculated such that if the tile has a smaller
start coordinate than the gene and the gene is on the + strand, distance will
be negative and calculated as min(0, end(tile) - start(gene))
}
\description{
ArchR-style gene activity scores are based on a weighted sum of each tile
according to the signed distance from the tile to a gene body. This function
calculates the signed distances according to ArchR's default parameters.
}
\details{
ArchR's tile distance algorithm works as follows
\enumerate{
\item Genes are extended 5kb upstream
\item Genes are linked to any tiles 1kb-100kb upstream + downstream, but tiles
beyond a neighboring gene are not considered
}

Note: assumes the 1-based, end inclusive coordinate convention used
by GRanges, so 500bp tiles run from bases 1-500, 501-1000, etc.
}
