% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneScores2.R
\name{gene_score_weights_archr}
\alias{gene_score_weights_archr}
\title{Calculate GeneActivityScore weights matrix}
\usage{
gene_score_weights_archr(
  genes,
  chromosome_sizes,
  blacklist = NULL,
  tile_size = 500,
  addArchRBug = FALSE
)
}
\arguments{
\item{genes}{GRanges object with gene start, end, and strand}

\item{chromosome_sizes}{\link{genomic-ranges} holding start and end coordinates for
each chromosome. See read_ucsc_chrom_sizes().}

\item{blacklist}{Blacklist ranges to exclude from gene calculations. \link{genomic-ranges}
object (list, data.frame, or GRanges)}

\item{tile_size}{Size of tiles to consider}

\item{addArchRBug}{Replicate ArchR bug in handling nested genes}
}
\value{
Weight matrix of dimension genes x tiles
}
\description{
Given a set of tile coordinates and distances returned by \code{gene_score_tiles()},
calculate a weight matrix of dimensions genes x tiles. This matrix can be
multiplied with a tile matrix to obtain ArchR-compatible gene activity scores.
}
