% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneScores.R
\name{geneScoreTilesArchR}
\alias{geneScoreTilesArchR}
\title{Calculate the tiles to use for GeneActivityScores, i.e. what counts as "core
gene" and what is the start and stop of regions that should be considered for
contributing to gene scores}
\usage{
geneScoreTilesArchR(genes, tile_size = 500, addArchRBug = FALSE)
}
\value{
GRanges object with one range per tile, with additional metadata
columns gene_idx (row index of the gene this tile corresponds to) and
distance.

Distance is a signed distance calculated such that if the tile has a smaller
start coordinate than the gene and the gene is on the + strand, distance will
be negative and calculated as min(0, end(tile) - start(gene))
}
\description{
Calculate the tiles to use for GeneActivityScores, i.e. what counts as "core
gene" and what is the start and stop of regions that should be considered for
contributing to gene scores
}
\details{
Note: assumes the 1-based, end inclusive coordinate convention used
by GRanges, so 500bp tiles run from bases 1-500, 501-1000, etc.
}
