% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneScores.R
\name{geneScoreWeightsArchR}
\alias{geneScoreWeightsArchR}
\title{Calculate GeneActivityScore weights matrix}
\usage{
geneScoreWeightsArchR(tiles, chr_sizes = GenomeInfoDb::seqlengths(tiles))
}
\arguments{
\item{tiles}{Tile coordinates as returned by \code{geneScoreTilesArchR}}

\item{chr_sizes}{Named vector of chromosome sizes with chromosmes in the order they are listed
in the tile matrix (used to calculate the index of each tile)}
}
\value{
Weight matrix of dimension genes x tiles
}
\description{
Given a set of tile coordinates and distances returned by \code{geneScoreTilesArchR},
calculate a weight matrix of dimensions genes x tiles. This matrix can be
multiplied with a tile matrix to obtain ArchR-compatible gene activity scores.
}
