% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPCells-package.R
\name{genomic-ranges-like}
\alias{genomic-ranges-like}
\title{Genomic range formats}
\description{
BPCells accepts a flexible set of genomic ranges-like objects as input, either
GRanges, data.frame, lists, or character vectors. These objects must specify chromosome, start,
and end coordinates along with optional metadata about each range.
With the exception of \code{GenomicRanges::GRanges} objects, BPCells assumes all
objects use a zero-based, end-exclusive coordinate system (see below for details).
\subsection{Valid Range-like objects}{

BPCells can interpret the following types as ranges:
\itemize{
\item \code{list()}, \code{data.frame()}, with columns:
\itemize{
\item \code{chr}: Character or factor of chromosome names
\item \code{start}: Start coordinates (0-based)
\item \code{end}: End coordinates (exclusive)
\item (optional) \code{strand}: \code{"+"}/\code{"-"} or \code{TRUE}/\code{FALSE} for pos/neg strand
\item (optional) Additional metadata as named \code{list} entries or \code{data.frame} columns
}
\item \code{GenomicRanges::GRanges}
\itemize{
\item \code{start(x)} is interpreted as a 1-based start coordinate
\item \code{end(x)} is interpreted as an inclusive end coordinate
\item \code{strand(x)}: \code{"*"} entries are interpeted as postive strand
\item (optional) \code{mcols(x)} holds additional metadata
}
\item \code{character}
\itemize{
\item Given in format \code{"chr1:1000-2000"} or \code{"chr1:1,000-2,000"}
\item Uses 0-based, end-exclusive coordinate system
\item Cannot be used for ranges where additional metadata is required
}
}
}

\subsection{Range coordinate systems}{

There are two main conventions for the coordinate systems:

\emph{\strong{One-based, end-inclusive ranges}}
\itemize{
\item The first base of a chromosome is numbered 1
\item The last base in a range is equal to the end coordinate
\item e.g. 1-5 describes the first 5 bases of the chromosome
\item Used in formats such as SAM, GTF
\item In BPCells, used when reading or writing \code{GenomicRanges::GRanges} objects
}

\emph{\strong{Zero-based, end-exclusive ranges}}
\itemize{
\item The first base of a chromosome is numbered 0
\item The last base in a range is one less than the end coordinate
\item e.g. 0-5 describes the first 5 bases of the chromosome
\item Used in formats such as BAM, BED
\item In BPCells, used for all other range objects
}
}
}
