% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{is_adjacency_matrix}
\alias{is_adjacency_matrix}
\title{Check if an input is a graph adjacency matrix.}
\usage{
is_adjacency_matrix(mat)
}
\value{
TRUE if the mat is a graph adjacency matrix, FALSE otherwise
}
\description{
Clustering function like \code{cluster_graph_leiden()} and \code{cluster_graph_louvain()} require a graph adjacency matrix as input.
We assume that any \code{dgCMatrix} that is square is a graph adjacency matrix.
}
\keyword{internal}
