% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{iterable_matrix_md5sum}
\alias{iterable_matrix_md5sum}
\title{Calculate IterableMatrix MD5 checksum as a string.}
\usage{
iterable_matrix_md5sum(matrix)
}
\description{
Calculate IterableMatrix MD5 checksum as a string.
}
