% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\docType{methods}
\name{chrNames,FragmentsTsv-method}
\alias{chrNames,FragmentsTsv-method}
\alias{cellNames,FragmentsTsv-method}
\alias{chrNames,UnpackedMemFragments-method}
\alias{cellNames,UnpackedMemFragments-method}
\alias{chrNames,PackedMemFragments-method}
\alias{cellNames,PackedMemFragments-method}
\alias{chrNames,FragmentsDir-method}
\alias{cellNames,FragmentsDir-method}
\alias{chrNames,FragmentsHDF5-method}
\alias{cellNames,FragmentsHDF5-method}
\alias{IterableFragments-misc-methods}
\alias{chrNames,ChrSelectName-method}
\alias{chrNames,ChrSelectIndex-method}
\alias{cellNames,CellSelectName-method}
\alias{cellNames,CellSelectIndex-method}
\alias{cellNames,CellMerge-method}
\alias{chrNames,ChrRename-method}
\alias{cellNames,CellRename-method}
\alias{cellNames,CellPrefix-method}
\alias{chrNames,MergeFragments-method}
\alias{cellNames,MergeFragments-method}
\title{IterableFragments subclass methods}
\usage{
\S4method{chrNames}{FragmentsTsv}(x)

\S4method{cellNames}{FragmentsTsv}(x)

\S4method{chrNames}{UnpackedMemFragments}(x)

\S4method{cellNames}{UnpackedMemFragments}(x)

\S4method{chrNames}{PackedMemFragments}(x)

\S4method{cellNames}{PackedMemFragments}(x)

\S4method{chrNames}{FragmentsDir}(x)

\S4method{cellNames}{FragmentsDir}(x)

\S4method{chrNames}{FragmentsHDF5}(x)

\S4method{cellNames}{FragmentsHDF5}(x)

\S4method{chrNames}{ChrSelectName}(x)

\S4method{chrNames}{ChrSelectIndex}(x)

\S4method{cellNames}{CellSelectName}(x)

\S4method{cellNames}{CellSelectIndex}(x)

\S4method{cellNames}{CellMerge}(x)

\S4method{chrNames}{ChrRename}(x)

\S4method{cellNames}{CellRename}(x)

\S4method{cellNames}{CellPrefix}(x)

\S4method{chrNames}{MergeFragments}(x)

\S4method{cellNames}{MergeFragments}(x)
}
\arguments{
\item{x}{An object inheriting from \code{IterableFragments}.}
}
\description{
Methods defined for classes that extend \code{IterableFragments}, providing access
to metadata or specialised behaviours for storage backends and selection
wrappers.
}
\section{Functions}{
\itemize{
\item \code{chrNames(FragmentsTsv)}: Get chromosome names for FragmentsTsv

\item \code{cellNames(FragmentsTsv)}: Get cell names for FragmentsTsv

\item \code{chrNames(UnpackedMemFragments)}: Get chromosome names for UnpackedMemFragments

\item \code{cellNames(UnpackedMemFragments)}: Get cell names for UnpackedMemFragments

\item \code{chrNames(PackedMemFragments)}: Get chromosome names for PackedMemFragments

\item \code{cellNames(PackedMemFragments)}: Get cell names for PackedMemFragments

\item \code{chrNames(FragmentsDir)}: Get chromosome names for FragmentsDir

\item \code{cellNames(FragmentsDir)}: Get cell names for FragmentsDir

\item \code{chrNames(FragmentsHDF5)}: Get chromosome names for FragmentsHDF5

\item \code{cellNames(FragmentsHDF5)}: Get cell names for FragmentsHDF5

\item \code{chrNames(ChrSelectName)}: Get chromosome names for ChrSelectName

\item \code{chrNames(ChrSelectIndex)}: Get chromosome names for ChrSelectIndex

\item \code{cellNames(CellSelectName)}: Get cell names for CellSelectName

\item \code{cellNames(CellSelectIndex)}: Get cell names for CellSelectIndex

\item \code{cellNames(CellMerge)}: Get cell names for CellMerge

\item \code{chrNames(ChrRename)}: Get chromosome names for ChrRename

\item \code{cellNames(CellRename)}: Get cell names for CellRename

\item \code{cellNames(CellPrefix)}: Get cell names for CellPrefix

\item \code{chrNames(MergeFragments)}: Get chromosome names for MergeFragments

\item \code{cellNames(MergeFragments)}: Get cell names for MergeFragments

}}
\keyword{internal}
