% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{IterableMatrix-methods-core}
\alias{IterableMatrix-methods-core}
\alias{matrix_type}
\alias{storage_order}
\alias{storage_order,IterableMatrix-method}
\alias{show,IterableMatrix-method}
\alias{t,IterableMatrix-method}
\alias{[,IterableMatrix,ANY,ANY,ANY-method}
\alias{[<-,IterableMatrix,ANY,ANY,ANY-method}
\alias{dimnames<-,IterableMatrix,list-method}
\alias{dimnames<-,IterableMatrix,NULL-method}
\title{IterableMatrix core methods}
\usage{
matrix_type(x)

storage_order(x)

\S4method{storage_order}{IterableMatrix}(x)

\S4method{show}{IterableMatrix}(object)

\S4method{t}{IterableMatrix}(x)

\S4method{[}{IterableMatrix,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{IterableMatrix,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{dimnames}{IterableMatrix,list}(x) <- value

\S4method{dimnames}{IterableMatrix,NULL}(x) <- value
}
\arguments{
\item{x}{IterableMatrix object}

\item{object}{IterableMatrix object}

\item{i}{Row indices or selection helpers.}

\item{j}{Column indices or selection helpers.}

\item{...}{Additional arguments passed to methods}

\item{drop}{Logical indicating whether to drop dimensions when subsetting.}

\item{value}{New dimnames (list of length 2, or NULL)}
}
\value{
\itemize{
\item \code{t()} Transposed object
}
}
\description{
Core operations for IterableMatrix objects: inspection, transpose, subsetting, and dimnames.
}
\section{Functions}{
\itemize{
\item \code{matrix_type()}: Get the matrix data type (mat_uint32_t, mat_float, or mat_double for now)

\item \code{storage_order()}: Get the matrix storage order ("row" or "col") (generic)

\item \code{storage_order(IterableMatrix)}: Get the matrix storage order ("row" or "col")

\item \code{show(IterableMatrix)}: Display an IterableMatrix

\item \code{t(IterableMatrix)}: Transpose an IterableMatrix

\item \code{x[i}: Subset an IterableMatrix

\item \code{`[`(x = IterableMatrix, i = ANY, j = ANY) <- value}: Assign into an IterableMatrix

\item \code{dimnames(x = IterableMatrix) <- value}: Set dimnames of an IterableMatrix, similar to base R \verb{dimnames<-()}

\item \code{dimnames(x = IterableMatrix) <- value}: Remove dimnames of an IterableMatrix

}}
\examples{
## Prep data
mat <- matrix(1:25, nrow = 5) \%>\% as("dgCMatrix")
mat
mat <- as(mat, "IterableMatrix")
mat


#######################################################################
## matrix_type() example
#######################################################################
matrix_type(mat)


#######################################################################
## storage_order() example
#######################################################################
storage_order(mat)


#######################################################################
## show() example
#######################################################################
show(mat)


#######################################################################
## t() example
#######################################################################
t(mat)


#' ####################################################################
## x[i, j, ..., drop = TRUE] example
#######################################################################
mat[1:2, 1:2]
#######################################################################
## x[i, j, ...] <- value example
#######################################################################
mat_changed <- mat
new_mat <- matrix(rep(2,4), nrow = 2) \%>\% as("IterableMatrix")
mat_changed[1:2, 1:2] <- new_mat
mat_changed \%>\% as("dgCMatrix")


}
