% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R, R/transforms.R
\name{IterableMatrix-methods-ops}
\alias{IterableMatrix-methods-ops}
\alias{\%*\%,IterableMatrix,matrix-method}
\alias{^,IterableMatrix,numeric-method}
\alias{<,numeric,IterableMatrix-method}
\alias{>,IterableMatrix,numeric-method}
\alias{<=,numeric,IterableMatrix-method}
\alias{>=,IterableMatrix,numeric-method}
\alias{round,IterableMatrix-method}
\alias{*,IterableMatrix,numeric-method}
\alias{+,IterableMatrix,numeric-method}
\alias{/,IterableMatrix,numeric-method}
\alias{-,IterableMatrix,numeric-method}
\title{IterableMatrix operations}
\usage{
\S4method{\%*\%}{IterableMatrix,matrix}(x, y)

\S4method{^}{IterableMatrix,numeric}(e1, e2)

\S4method{<}{numeric,IterableMatrix}(e1, e2)

\S4method{>}{IterableMatrix,numeric}(e1, e2)

\S4method{<=}{numeric,IterableMatrix}(e1, e2)

\S4method{>=}{IterableMatrix,numeric}(e1, e2)

\S4method{round}{IterableMatrix}(x, digits = 0)

\S4method{*}{IterableMatrix,numeric}(e1, e2)

\S4method{+}{IterableMatrix,numeric}(e1, e2)

\S4method{/}{IterableMatrix,numeric}(e1, e2)

\S4method{-}{IterableMatrix,numeric}(e1, e2)
}
\arguments{
\item{x}{IterableMatrix object}

\item{y}{matrix}

\item{e1}{First element of comparison}

\item{e2}{Second element of comparison}

\item{digits}{Number of decimal places for rounding}
}
\value{
\itemize{
\item \code{x \%*\% y}: dense matrix result
}
}
\description{
Matrix multiplication, arithmetic, and comparison operations for IterableMatrix objects.
}
\section{Functions}{
\itemize{
\item \code{x \%*\% y}: Multiply by a dense matrix

\item \code{e1^e2}: Calculate x^y (elementwise; y > 0)

\item \code{e1 < e2}: Perform numeric < matrix comparison (numeric left)

\item \code{e1 > e2}: Perform matrix > numeric comparison (IterableMatrix left)

\item \code{e1 <= e2}: Perform numeric <= matrix comparison (numeric left)

\item \code{e1 >= e2}: Perform matrix >= numeric comparison (IterableMatrix left)

\item \code{round(IterableMatrix)}: round to nearest integer (digits must be 0)

\item \code{e1 * e2}: Multiply by a constant, or multiply rows by a vector length nrow(mat)

\item \code{e1 + e2}: Add a constant, or row-wise addition with a vector length nrow(mat)

\item \code{e1 / e2}: Divide by a constant, or divide rows by a vector length nrow(mat)

\item \code{e1 - e2}: Subtract a constant, or row-wise subtraction with a vector length nrow(mat)

}}
\examples{
## Prep data
mat <- matrix(1:25, nrow = 5) \%>\% as("dgCMatrix")
mat
mat <- as(mat, "IterableMatrix")
mat

#######################################################################
## `x \%*\% y` example
#######################################################################
mat \%*\% as(matrix(1:50, nrow = 5), "dgCMatrix")


5 < mat


#######################################################################
## `e1 > e2` example
#######################################################################
mat > 5


#######################################################################
## `e1 <= e2` example
#######################################################################
5 <= mat


#######################################################################
## `e1 >= e2` example
#######################################################################
mat >= 5


#######################################################################
## round() example
#######################################################################
round(mat)


#######################################################################
## `e1 * e2` example
#######################################################################
## Multiplying by a constant
mat * 5

## Multiplying by a vector of length `nrow(mat)`
mat * 1:nrow(mat)


#######################################################################
## `e1 + e2` example
#######################################################################
## Add by a constant
mat + 5

## Adding row-wise by a vector of length `nrow(mat)`
mat + 1:nrow(mat)


#######################################################################
## `e1 / e2` example
#######################################################################
## Divide by a constant
mat / 5

## Divide by a vector of length `nrow(mat)`
mat / 1:nrow(mat)


#######################################################################
## `e1 - e2` example
#######################################################################
## Subtracting by a constant
mat - 5

## Subtracting by a vector of length `nrow(mat)`
mat - 1:nrow(mat)


}
