% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R, R/transforms.R
\docType{methods}
\name{IterableMatrix-misc-methods}
\alias{IterableMatrix-misc-methods}
\alias{matrix_type,MatrixMultiply-method}
\alias{[,MatrixMultiply,ANY,ANY,ANY-method}
\alias{matrix_type,MatrixMask-method}
\alias{matrix_type,MatrixRankTransform-method}
\alias{matrix_type,MatrixSubset-method}
\alias{[,MatrixSubset,ANY,ANY,ANY-method}
\alias{matrix_type,RenameDims-method}
\alias{[,RenameDims,ANY,ANY,ANY-method}
\alias{matrix_type,RowBindMatrices-method}
\alias{matrix_type,ColBindMatrices-method}
\alias{[,RowBindMatrices,ANY,ANY,ANY-method}
\alias{[,ColBindMatrices,ANY,ANY,ANY-method}
\alias{matrix_type,PackedMatrixMem_uint32_t-method}
\alias{matrix_type,PackedMatrixMem_float-method}
\alias{matrix_type,PackedMatrixMem_double-method}
\alias{matrix_type,UnpackedMatrixMem_uint32_t-method}
\alias{matrix_type,UnpackedMatrixMem_float-method}
\alias{matrix_type,UnpackedMatrixMem_double-method}
\alias{matrix_type,MatrixDir-method}
\alias{matrix_type,EXPERIMENTAL_MatrixDirCompressedCol-method}
\alias{matrix_type,MatrixH5-method}
\alias{matrix_type,10xMatrixH5-method}
\alias{matrix_type,AnnDataMatrixH5-method}
\alias{matrix_type,PeakMatrix-method}
\alias{[,PeakMatrix,ANY,ANY,ANY-method}
\alias{matrix_type,TileMatrix-method}
\alias{[,TileMatrix,ANY,ANY,ANY-method}
\alias{matrix_type,ConvertMatrixType-method}
\alias{[,ConvertMatrixType,ANY,ANY,ANY-method}
\alias{matrix_type,Iterable_dgCMatrix_wrapper-method}
\alias{matrix_type,TransformedMatrix-method}
\alias{[,TransformedMatrix,ANY,ANY,ANY-method}
\alias{*,TransformScaleShift,numeric-method}
\alias{+,TransformScaleShift,numeric-method}
\alias{*,numeric,TransformScaleShift-method}
\alias{+,numeric,TransformScaleShift-method}
\title{IterableMatrix subclass methods}
\usage{
\S4method{matrix_type}{MatrixMultiply}(x)

\S4method{[}{MatrixMultiply,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{matrix_type}{MatrixMask}(x)

\S4method{matrix_type}{MatrixRankTransform}(x)

\S4method{matrix_type}{MatrixSubset}(x)

\S4method{[}{MatrixSubset,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{matrix_type}{RenameDims}(x)

\S4method{[}{RenameDims,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{matrix_type}{RowBindMatrices}(x)

\S4method{matrix_type}{ColBindMatrices}(x)

\S4method{[}{RowBindMatrices,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{ColBindMatrices,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{matrix_type}{PackedMatrixMem_uint32_t}(x)

\S4method{matrix_type}{PackedMatrixMem_float}(x)

\S4method{matrix_type}{PackedMatrixMem_double}(x)

\S4method{matrix_type}{UnpackedMatrixMem_uint32_t}(x)

\S4method{matrix_type}{UnpackedMatrixMem_float}(x)

\S4method{matrix_type}{UnpackedMatrixMem_double}(x)

\S4method{matrix_type}{MatrixDir}(x)

\S4method{matrix_type}{EXPERIMENTAL_MatrixDirCompressedCol}(x)

\S4method{matrix_type}{MatrixH5}(x)

\S4method{matrix_type}{10xMatrixH5}(x)

\S4method{matrix_type}{AnnDataMatrixH5}(x)

\S4method{matrix_type}{PeakMatrix}(x)

\S4method{[}{PeakMatrix,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{matrix_type}{TileMatrix}(x)

\S4method{[}{TileMatrix,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{matrix_type}{ConvertMatrixType}(x)

\S4method{[}{ConvertMatrixType,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{matrix_type}{Iterable_dgCMatrix_wrapper}(x)

\S4method{matrix_type}{TransformedMatrix}(x)

\S4method{[}{TransformedMatrix,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{*}{TransformScaleShift,numeric}(e1, e2)

\S4method{+}{TransformScaleShift,numeric}(e1, e2)

\S4method{*}{numeric,TransformScaleShift}(e1, e2)

\S4method{+}{numeric,TransformScaleShift}(e1, e2)
}
\arguments{
\item{x}{An object inheriting from \code{IterableMatrix}.}

\item{i}{Row indices or selection helpers.}

\item{j}{Column indices or selection helpers.}

\item{...}{Additional arguments passed through the call.}

\item{drop}{Logical indicating whether to drop dimensions (for subsetting).}

\item{e1}{Left operand for binary operations.}

\item{e2}{Right operand for binary operations.}
}
\description{
Methods for classes that extend \code{IterableMatrix} but are not dispatched
directly on the base class. These are typically helper objects that wrap
another matrix or alter behaviour (e.g., concatenation, on-disk access).
}
\section{Functions}{
\itemize{
\item \code{matrix_type(MatrixMultiply)}: Matrix data type for MatrixMultiply objects

\item \code{x[i}: Subset MatrixMultiply results

\item \code{matrix_type(MatrixMask)}: Matrix data type for MatrixMask objects

\item \code{matrix_type(MatrixRankTransform)}: Matrix data type for MatrixRankTransform objects

\item \code{matrix_type(MatrixSubset)}: Matrix data type for MatrixSubset objects

\item \code{x[i}: Subset MatrixSubset transforms

\item \code{matrix_type(RenameDims)}: Matrix data type for RenameDims objects

\item \code{x[i}: Subset RenameDims transforms

\item \code{matrix_type(RowBindMatrices)}: Matrix data type for RowBindMatrices objects

\item \code{matrix_type(ColBindMatrices)}: Matrix data type for ColBindMatrices objects

\item \code{x[i}: Subset RowBindMatrices transforms

\item \code{x[i}: Subset ColBindMatrices transforms

\item \code{matrix_type(PackedMatrixMem_uint32_t)}: Matrix data type for PackedMatrixMem_uint32_t objects

\item \code{matrix_type(PackedMatrixMem_float)}: Matrix data type for PackedMatrixMem_float objects

\item \code{matrix_type(PackedMatrixMem_double)}: Matrix data type for PackedMatrixMem_double objects

\item \code{matrix_type(UnpackedMatrixMem_uint32_t)}: Matrix data type for UnpackedMatrixMem_uint32_t objects

\item \code{matrix_type(UnpackedMatrixMem_float)}: Matrix data type for UnpackedMatrixMem_float objects

\item \code{matrix_type(UnpackedMatrixMem_double)}: Matrix data type for UnpackedMatrixMem_double objects

\item \code{matrix_type(MatrixDir)}: Matrix data type for MatrixDir objects

\item \code{matrix_type(EXPERIMENTAL_MatrixDirCompressedCol)}: Matrix data type for EXPERIMENTAL_MatrixDirCompressedCol objects

\item \code{matrix_type(MatrixH5)}: Matrix data type for MatrixH5 objects

\item \code{matrix_type(`10xMatrixH5`)}: Matrix data type for 10xMatrixH5 objects

\item \code{matrix_type(AnnDataMatrixH5)}: Matrix data type for AnnDataMatrixH5 objects

\item \code{matrix_type(PeakMatrix)}: Matrix data type for PeakMatrix objects

\item \code{x[i}: Subset a PeakMatrix

\item \code{matrix_type(TileMatrix)}: Matrix data type for TileMatrix objects

\item \code{x[i}: Subset a TileMatrix

\item \code{matrix_type(ConvertMatrixType)}: Matrix data type for ConvertMatrixType objects

\item \code{x[i}: Subset ConvertMatrixType transforms

\item \code{matrix_type(Iterable_dgCMatrix_wrapper)}: Matrix data type for Iterable_dgCMatrix_wrapper objects

\item \code{matrix_type(TransformedMatrix)}: Matrix data type for TransformedMatrix objects

\item \code{x[i}: Subset TransformedMatrix results

\item \code{e1 * e2}: Scale TransformScaleShift results by numeric values

\item \code{e1 + e2}: Shift TransformScaleShift results by numeric values

\item \code{e1 * e2}: Apply numeric scaling on the left to TransformScaleShift results

\item \code{e1 + e2}: Add TransformScaleShift results to numeric values (numeric left operand)

}}
\keyword{internal}
