% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{iterate_matrix}
\alias{iterate_matrix}
\title{Get a wrapped pointer to the iterable matrix, in the form of an XPtrList (see fragments.R)}
\usage{
iterate_matrix(x)
}
\description{
Get a wrapped pointer to the iterable matrix, in the form of an XPtrList (see fragments.R)
}
