% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac_utils.R
\name{length_distribution}
\alias{length_distribution}
\title{Count fragments by size}
\usage{
length_distribution(fragments)
}
\arguments{
\item{fragments}{Fragments object}
}
\value{
Numeric vector where index i contans the number of length-i fragments
}
\description{
Count fragments by size
}
