% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{linear_regress_out}
\alias{linear_regress_out}
\title{Linear Regression}
\usage{
linear_regress_out(mat, latent_data = NULL)
}
\arguments{
\item{mat}{IterableMatrix (raw counts)}

\item{latent_data}{Extra data to regress out, should be cells x latent \code{data.frame}}
}
\value{
IterableMatrix
}
\description{
Calculate the residuals from a linear model fit.
}
