% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\docType{class}
\name{LinearOperator-class}
\alias{LinearOperator-class}
\title{LinearOperator class for performing sparse matrix - dense vector products, e.g.
for downstream matrix solvers. This is to avoid having to call iterate_matrix
repeatedly for a possible efficiency gain}
\description{
LinearOperator class for performing sparse matrix - dense vector products, e.g.
for downstream matrix solvers. This is to avoid having to call iterate_matrix
repeatedly for a possible efficiency gain
}
