% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\docType{methods}
\name{LinearOperator-math}
\alias{LinearOperator-math}
\alias{\%*\%,LinearOperator,matrix-method}
\alias{\%*\%,matrix,LinearOperator-method}
\alias{\%*\%,LinearOperator,numeric-method}
\alias{\%*\%,numeric,LinearOperator-method}
\title{LinearOperator multiplication helpers}
\usage{
\S4method{\%*\%}{LinearOperator,matrix}(x, y)

\S4method{\%*\%}{matrix,LinearOperator}(x, y)

\S4method{\%*\%}{LinearOperator,numeric}(x, y)

\S4method{\%*\%}{numeric,LinearOperator}(x, y)
}
\arguments{
\item{x}{Left operand.}

\item{y}{Right operand.}
}
\description{
Methods enabling \verb{\\\%*\%} between \code{LinearOperator} objects and dense matrices or numeric vectors.
}
\section{Functions}{
\itemize{
\item \code{x \%*\% y}: Multiply a LinearOperator by a dense matrix

\item \code{x \%*\% y}: Multiply a dense matrix by a LinearOperator

\item \code{x \%*\% y}: Multiply a LinearOperator by a numeric vector

\item \code{x \%*\% y}: Multiply a numeric vector by a LinearOperator

}}
\keyword{internal}
