% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_transformers.R
\docType{class}
\name{LSITransformer}
\alias{LSITransformer}
\title{Perform latent semantic indexing (LSI) on a matrix.
TODO: Add more details from when upstream LSI PR is reviewed.}
\usage{
LSITransformer(
  z_score_norm = TRUE,
  n_dimensions = 50L,
  scale_factor = 10000L,
  threads = 1L
)
}
\description{
Perform latent semantic indexing (LSI) on a matrix.
TODO: Add more details from when upstream LSI PR is reviewed.

Create a new LSITransformer object
}
