% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{mask_matrix}
\alias{mask_matrix}
\title{Mask matrix entries to zero
Set matrix entries to zero given a mask matrix of the
same dimensions. Normally, non-zero values in the mask
will set the matrix entry to zero. If inverted, zero
values in the mask matrix will set the matrix entry to zero.}
\usage{
mask_matrix(mat, mask, invert = FALSE)
}
\arguments{
\item{mat}{Data matrix (IterableMatrix)}

\item{mask}{Mask matrix (IterableMatrix or dgCMatrix)}
}
\description{
Mask matrix entries to zero
Set matrix entries to zero given a mask matrix of the
same dimensions. Normally, non-zero values in the mask
will set the matrix entry to zero. If inverted, zero
values in the mask matrix will set the matrix entry to zero.
}
\keyword{internal}
