% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix_is_transform}
\alias{matrix_is_transform}
\title{Return boolean whether the matrix is a transform (i.e. is loading data from another
matrix R object rather than a data source it owns). This is used primarily to know when it is safe
to clear dimnames from intermediate transformed matrices. C++ relies on the base matrices (non-transform) to have
dimnames, while R relies on the outermost matrix (transform) to have dimnames.}
\usage{
matrix_is_transform(x)
}
\description{
Return boolean whether the matrix is a transform (i.e. is loading data from another
matrix R object rather than a data source it owns). This is used primarily to know when it is safe
to clear dimnames from intermediate transformed matrices. C++ relies on the base matrices (non-transform) to have
dimnames, while R relies on the outermost matrix (transform) to have dimnames.
}
