% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell_utils.R
\name{matrix_quantile_per_cell}
\alias{matrix_quantile_per_cell}
\title{Find the nth quantile value of each cell in a matrix.}
\usage{
matrix_quantile_per_cell(mat, quantile = 0.99)
}
\arguments{
\item{mat}{IterableMatrix object of dimensions features x cells}

\item{quantile}{(numeric) quantile value to be found from each cell, between \link{0, 1}.}
}
\value{
(Numeric) cell quantile values with number of entries equal to number of colums in the matrix.
}
\description{
Find the nth quantile value of each cell in a matrix.
}
