% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\docType{methods}
\name{matrix_r_conversion-misc}
\alias{matrix_r_conversion-misc}
\alias{coerce,dgCMatrix,IterableMatrix-method}
\alias{coerce,IterableMatrix,dgCMatrix-method}
\alias{coerce,matrix,IterableMatrix-method}
\alias{coerce,IterableMatrix,matrix-method}
\title{Convert between BPCells matrices and R objects misc.}
\usage{
\S4method{coerce}{dgCMatrix,IterableMatrix}(from, to, ...)
\S4method{coerce}{IterableMatrix,dgCMatrix}(from, to, ...)
\S4method{coerce}{matrix,IterableMatrix}(from, to, ...)
\S4method{coerce}{IterableMatrix,matrix}(from, to, ...)
}
\arguments{
\item{from}{Object supplied to \code{base::coerce()} (typically generated by \code{as()})}

\item{to}{Target class name for coercion}

\item{...}{Additional arguments passed to methods}
}
\description{
See \code{matrix_R_conversion} for main page.
BPCells matrices can be interconverted with Matrix package \code{dgCMatrix} sparse matrices,
as well as base R dense matrices (though this may result in high memory usage for large matrices).
}
\keyword{internal}
