% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix_R_conversion_coercions}
\alias{matrix_R_conversion_coercions}
\alias{as.matrix.IterableMatrix}
\alias{as.matrix,IterableMatrix-method}
\alias{coerce,dgCMatrix,IterableMatrix-method}
\alias{coerce,IterableMatrix,dgCMatrix-method}
\alias{coerce,matrix,IterableMatrix-method}
\alias{coerce,IterableMatrix,matrix-method}
\alias{matrix_R_conversion}
\title{Convert between BPCells matrix and R objects.}
\usage{
\method{as.matrix}{IterableMatrix}(x, ...)
\S4method{coerce}{dgCMatrix,IterableMatrix}(from, to, ...)
\S4method{coerce}{IterableMatrix,dgCMatrix}(from, to, ...)
\S4method{coerce}{matrix,IterableMatrix}(from, to, ...)
\S4method{coerce}{IterableMatrix,matrix}(from, to, ...)
}
\arguments{
\item{x}{Matrix object to convert}

\item{dgCMatrix}{Sparse matrix in dgCMatrix format}

\item{IterableMatrix}{IterableMatrix object}

\item{...}{Additional arguments passed to methods}

\item{from}{Object supplied to \code{base::coerce()} (typically generated by \code{as()})}

\item{to}{Target class name for coercion}
}
\description{
BPCells matrices can be interconverted with Matrix package
dgCMatrix sparse matrices, as well as base R
dense matrices (though this may result in high memory usage for large matrices)
}
\details{
Use base R's \code{as()} to convert between BPCells matrices and
\code{dgCMatrix}/\code{matrix} representations, while \code{as.matrix()} materialises
dense matrices directly when needed.
}
\examples{
mat <- get_demo_mat()[1:2, 1:2]
mat


#######################################################################
## as(bpcells_mat, "dgCMatrix") example
#######################################################################
mat_dgc <- as(mat, "dgCMatrix")
mat_dgc


## as.matrix(bpcells_mat) example
as.matrix(mat)

## Alternatively, can also use function as()
as(mat, "matrix")

#######################################################################
## as(dgc_mat, "IterableMatrix") example
#######################################################################
as(mat_dgc, "IterableMatrix")


}
