% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix_type}
\alias{matrix_type}
\title{Get the matrix data type (mat_uint32_t, mat_float, or mat_double for now)}
\usage{
matrix_type(x)
}
\description{
Get the matrix data type (mat_uint32_t, mat_float, or mat_double for now)
}
