% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrixStats}
\alias{matrixStats}
\title{Calculate matrix stats}
\usage{
matrixStats(
  matrix,
  row_stats = c("none", "nonzero", "mean", "variance"),
  col_stats = c("none", "nonzero", "mean", "variance")
)
}
\arguments{
\item{matrix}{Input matrix object}

\item{row_stats}{Which row statistics to compute}

\item{col_stats}{Which col statistics to compute}
}
\value{
List of row_stats: matrix of n_stats x n_rows,
col_stats: matrix of n_stats x n_cols
}
\description{
Calculate matrix stats
}
\details{
The statistics will be calculated in a single pass over the matrix,
so this method is desirable to use for efficiency purposes compared to
the more standard rowMeans or colMeans if multiple statistics are needed.
If variance is calculated, then mean and nonzero count will be included in the
output, and if mean is calculated then nonzero count will be included in the output.
}
