% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{merge_dimnames}
\alias{merge_dimnames}
\title{Helper function for rbind/cbind merging dimnames}
\usage{
merge_dimnames(x, y, warning_prefix, dim_type)
}
\description{
Helper function for rbind/cbind merging dimnames
}
\keyword{internal}
