% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac_utils.R
\name{merge_peaks_iterative}
\alias{merge_peaks_iterative}
\title{Merge peaks}
\usage{
merge_peaks_iterative(peaks)
}
\arguments{
\item{peaks}{\code{data.frame}, \code{tibble}, or \code{list} ranges object. Must be ordered by
priority and have columns chr, start, end.}
}
\value{
`tibble::tibble()`` with a nonoverlapping subset of the rows in peaks. All metadata
columns are preserved
}
\description{
Merge peaks according to ArchR's iterative merging algorithm. More
details here: \url{https://www.archrproject.com/bookdown/the-iterative-overlap-peak-merging-procedure.html}
}
\details{
Properties of merged peaks:
\itemize{
\item No peaks in the merged set overlap
\item Peaks are prioritized according to their order in the original input
\item The output peaks are a subset of the input peaks, with no peak boundaries
changed
}
}
