% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{min_scalar}
\alias{min_scalar}
\alias{min_by_row}
\alias{min_by_col}
\title{Elementwise minimum}
\usage{
min_scalar(mat, val)

min_by_row(mat, vals)

min_by_col(mat, vals)
}
\arguments{
\item{mat}{IterableMatrix}

\item{val}{Single positive numeric value}
}
\value{
IterableMatrix
}
\description{
\strong{min_scalar}: Take minumum with a global constant

\strong{min_by_row}: Take the minimum with a per-row constant

\strong{min_by_col}: Take the minimum with a per-col constant
}
\details{
Take the minimum value of a matrix with a per-row, per-col, or global
constant. This constant must be >0 to preserve sparsity of the matrix.
This has the effect of capping the maximum value in the matrix.
}
