% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{min_scalar}
\alias{min_scalar}
\title{Take the elementwise min with a constant}
\usage{
min_scalar(mat, val)
}
\arguments{
\item{mat}{IterableMatrix}

\item{val}{Single positive numeric value}
}
\value{
IterableMatrix
}
\description{
Take the elementwise minimum with a positive value.
This has the effect of capping the maximum value in the matrix
}
