% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{normalize_log}
\alias{normalize_log}
\title{Normalize a matrix using log normalization}
\usage{
normalize_log(mat, scale_factor = 10000, add_one = TRUE)
}
\arguments{
\item{mat}{(IterableMatrix) Matrix to normalize}

\item{scale_factor}{(numeric) Scale factor to multiply matrix by for log normalization}

\item{add_one}{(logical) Add one to the matrix before log normalization}
}
\value{
log normalized matrix.
}
\description{
Normalize a matrix using log normalization
}
