% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{normalize_tfidf}
\alias{normalize_tfidf}
\title{Normalize a `(features x cells)`` matrix using term frequency-inverse document frequency}
\usage{
normalize_tfidf(mat, feature_means = NULL, threads = 1L)
}
\arguments{
\item{mat}{(IterableMatrix) to normalize}

\item{feature_means}{(numeric) Means of the features to normalize by.  If no names are provided, then
each numeric value is assumed to correspond to the feature mean for the corresponding row of the matrix.
Else, map each feature name to its mean value.}
}
\value{
tf-idf normalized matrix.
}
\description{
Normalize a `(features x cells)`` matrix using term frequency-inverse document frequency
}
