% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorChecking.R
\name{normalize_unique_file_names}
\alias{normalize_unique_file_names}
\title{Adjust a set of (unique) potential file names to not include any
invalid characters.}
\usage{
normalize_unique_file_names(names, replacement = "_")
}
\description{
Adjust a set of (unique) potential file names to not include any
invalid characters.
}
\keyword{internal}
