% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{normalized_dimnames}
\alias{normalized_dimnames}
\title{Helper function to set dimnames to NULL instead of 0-length character vectors}
\usage{
normalized_dimnames(row_names, col_names)
}
\description{
Helper function to set dimnames to NULL instead of 0-length character vectors
}
\keyword{internal}
