% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacQC.R
\name{nucleosome_counts}
\alias{nucleosome_counts}
\title{Count fragments by nucleosomal size}
\usage{
nucleosome_counts(fragments, nucleosome_width = 147)
}
\arguments{
\item{fragments}{Fragments object}

\item{nucleosome_width}{Integer cutoff to use as nucleosome width}
}
\value{
List with names subNucleosomal, monoNucleosomal, multiNucleosomal containing the
count vectors of fragments in each class per cell.
}
\description{
Count fragments by nucleosomal size
}
