% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{open_fragments_10x}
\alias{open_fragments_10x}
\title{Read a 10x fragments file}
\usage{
open_fragments_10x(path, comment = "#", end_inclusive = TRUE)
}
\arguments{
\item{path}{Path of 10x fragments file}

\item{comment}{Skip lines at beginning of file which start with comment}

\item{end_inclusive}{Whether the end coordinate of the bed is inclusive -- i.e. there was an
insertion at the end coordinate rather than the base before the end coordinate. This is the
10x default, though it's not quite standard for the bed file format.}
}
\value{
10x fragments file object
}
\description{
Read a 10x fragments file
}
\details{
Note: no disk operations will take place until the fragments are used in a function
}
