% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{open_fragments_dir}
\alias{open_fragments_dir}
\title{Open an existing fragments object written with \link{write_fragments_dir}}
\usage{
open_fragments_dir(dir, buffer_size = 1024L)
}
\arguments{
\item{dir}{Directory to read the data from}

\item{buffer_size}{For performance tuning only. The number of fragments to be buffered
in memory for each read}
}
\value{
FragmentsDir object
}
\description{
Open an existing fragments object written with \link{write_fragments_dir}
}
