% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{open_fragments_hdf5}
\alias{open_fragments_hdf5}
\title{Read a fragments object from an hdf5 file}
\usage{
open_fragments_hdf5(path, group = "fragments", buffer_size = 16384L)
}
\arguments{
\item{path}{Path to the hdf5 file on disk}

\item{group}{The group within the hdf5 file to write the data to. If writing
to an existing hdf5 file this group must not already be in use}

\item{buffer_size}{For performance tuning only. The number of items to be bufferred
in memory before calling reads from disk.}
}
\value{
FragmentsHDF5 object
}
\description{
Read a fragments object from an hdf5 file
}
