% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{open_matrix_10x_hdf5}
\alias{open_matrix_10x_hdf5}
\title{Read a sparse integer matrix from a 10x formatted hdf5 feature matrix file}
\usage{
open_matrix_10x_hdf5(path, buffer_size = 16384L)
}
\arguments{
\item{path}{Path to the hdf5 file on disk}

\item{buffer_size}{For performance tuning only. The number of items to be buffered
in memory before calling writes to disk.}
}
\value{
10xMatrixH5 object
}
\description{
Read a sparse integer matrix from a 10x formatted hdf5 feature matrix file
}
\details{
The 10x format makes use of gzip compression for the matrix data,
which can slow down read performance. Consider writing into another format
if the read performance is important to you.
}
