% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{open_matrix_anndata_hdf5}
\alias{open_matrix_anndata_hdf5}
\title{Read AnnData matrix}
\usage{
open_matrix_anndata_hdf5(path, group = "X", buffer_size = 16384L)
}
\arguments{
\item{path}{Path to the hdf5 file on disk}

\item{group}{The group within the hdf5 file to write the data to. If writing
to an existing hdf5 file this group must not already be in use}

\item{buffer_size}{For performance tuning only. The number of items to be buffered
in memory before calling writes to disk.}
}
\value{
AnnDataMatrixH5 object, with cells as the columns.
}
\description{
Read a sparse integer matrix from an anndata matrix in an hdf5 file.
}
\details{
Since AnnData stores RNA matrices as cells x genes, whereas BPCells
stores RNA matrices as genes x cells, the returned matrix will be transposed
relative to the native AnnData matrix.
}
