% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{open_matrix_dir}
\alias{open_matrix_dir}
\title{Open an existing integer sparse matrix object written with \link{write_matrix_dir}}
\usage{
open_matrix_dir(dir, buffer_size = 8192L)
}
\arguments{
\item{dir}{Directory to read the data from}

\item{buffer_size}{For performance tuning only. The number of items to be buffered
in memory before calling writes to disk.}
}
\description{
Open an existing integer sparse matrix object written with \link{write_matrix_dir}
}
