% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{open_matrix_hdf5}
\alias{open_matrix_hdf5}
\title{Read a sparse integer matrix from an hdf5 file}
\usage{
open_matrix_hdf5(path, group, buffer_size = 16384L)
}
\arguments{
\item{path}{Path to the hdf5 file on disk}

\item{group}{The group within the hdf5 file to write the data to. If writing
to an existing hdf5 file this group must not already be in use}

\item{buffer_size}{For performance tuning only. The number of items to be buffered
in memory before calling writes to disk.}
}
\value{
MatrixH5 object
}
\description{
Read a sparse integer matrix from an hdf5 file
}
