% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_ranges}
\alias{order_ranges}
\title{Get end-sorted ordering for genome ranges}
\usage{
order_ranges(ranges, chr_levels)
}
\arguments{
\item{ranges}{List, GRanges, or data.frame object. Must have start, end, and chr defined}

\item{chr_levels}{Ordering of chromosome names}
}
\value{
Numeric vector analagous to the \code{order} function. Provides an index
selection that will reorder the input ranges to be sorted by chr, end, start
}
\description{
Use this function to order regioins prior to calling peakMatrix or tileMatrix.
}
