% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{parallel_split}
\alias{parallel_split}
\title{Prepare a matrix for multi-threaded operation}
\usage{
parallel_split(mat, threads, chunks = threads)
}
\arguments{
\item{mat}{IterableMatrix}

\item{threads}{Number of execution threads}

\item{chunks}{Number of chunks to use (>= threads)}
}
\value{
IterableMatrix which will perform certain operations in parallel
}
\description{
Transforms a matrix such that \code{matrix_stats} or matrix multiplies with
a vector/dense matrix will be evaluated in parallel. This only speeds up
those specific operations, not reading or writing the matrix in general.
The parallelism is not guaranteed to work if additional operations are
applied after the parallel split.
}
\keyword{internal}
