% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partial_apply}
\alias{partial_apply}
\title{Create partial function calls}
\usage{
partial_apply(f, ..., .overwrite = FALSE)
}
\arguments{
\item{f}{A function}

\item{...}{Named arguments to \code{f}}

\item{.overwrite}{(bool) If \code{f} is already an output from
\code{partial_apply()}, whether parameter re-definitions should
be ignored or overwrite the existing definitions}
}
\value{
A \code{bpcells_partial} object (a function with some extra attributes)
}
\description{
Specify some but not all arguments to a function.
}
\keyword{internal}
