% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_pipelines.R
\docType{class}
\name{Pipeline}
\alias{Pipeline}
\title{S4 class for combining multiple pipeline steps into a single pipeline}
\usage{
Pipeline(steps = list())
}
\arguments{
\item{steps}{A list of ordered steps to be executed in the pipeline.}
}
\value{
A new Pipeline object.
}
\description{
S4 class for combining multiple pipeline steps into a single pipeline

Return a new Pipeline object.
}
\details{
Creating a pipeline object can be done by passing a list of pipeline steps to the constructor.
Creation only expects that all steps make logical sense.  i.e., the final step can be either an Estimator or a Transformer,
but each intermediate step cannot be an Estimator.
}
\section{Slots}{

\describe{
\item{\code{steps}}{(list) List of pipeline steps to execute in order}
}}

