% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_project_predict.R
\docType{class}
\name{PipelineBase}
\alias{PipelineBase}
\title{Pipeline Base Class}
\description{
Pipeline Base Class
}
\section{Slots}{

\describe{
\item{\code{fitted}}{(logical) Whether the pipeline has been fitted}
}}

