% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_pipelines.R
\docType{class}
\name{PipelineStep}
\alias{PipelineStep}
\title{PipelineBase representing a single step within a pipeline}
\description{
PipelineBase representing a single step within a pipeline
}
\section{Slots}{

\describe{
\item{\code{step_name}}{(character) Name of the step}

\item{\code{fitted}}{(logical) Whether the pipeline has been fitted}
}}

