% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_read_count_knee}
\alias{plot_read_count_knee}
\title{Knee plot of single cell read counts}
\usage{
plot_read_count_knee(
  read_counts,
  cutoff = NULL,
  return_data = FALSE,
  apply_styling = TRUE
)
}
\arguments{
\item{read_counts}{Vector of read counts per cell}

\item{cutoff}{(optional) Read cutoff to mark on the plot}

\item{return_data}{If true, return data from just before plotting rather than a plot.}

\item{apply_styling}{If false, return a plot without pretty styling applied}
}
\value{
ggplot2 plot object
}
\description{
Plots read count rank vs. number of reads on a log-log scale.
}
\details{
Performs logarithmic downsampling to reduce the number of points plotted
}
\examples{
## Prep data
mat <- get_demo_mat(filter_qc = FALSE, subset = FALSE)
reads_per_cell <- colSums(mat)

# Render knee plot
plot_read_count_knee(reads_per_cell, cutoff = 1e3)
}
