% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_project_predict.R
\name{predict(PipelineBase,IterableMatrix)}
\alias{predict(PipelineBase,IterableMatrix)}
\title{Predict the output data using a fitted pipeline}
\usage{
predict(object, x, ...)
}
\arguments{
\item{object}{(PipelineBase) The fitted pipeline object}

\item{x}{(IterableMatrix) Input data to be predicted}
}
\value{
Predicted output data
}
\description{
Predict the output data using a fitted pipeline
}
