% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{prefix_cell_names}
\alias{prefix_cell_names}
\title{Rename cells by adding a prefix to the names (e.g. sample name)}
\usage{
prefix_cell_names(
  fragments,
  prefix,
  invert_selection = FALSE,
  zero_based_coords = TRUE
)
}
\arguments{
\item{fragments}{Input fragments object.}

\item{prefix}{String to add as the prefix}
}
\value{
Fragments object with prefixed names
}
\description{
Rename cells by adding a prefix to the names (e.g. sample name)
}
