% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac_utils.R
\name{prep_macs_inputs}
\alias{prep_macs_inputs}
\title{Call MACS2/3 peaks using fragments. UNFINISHED.}
\usage{
prep_macs_inputs(
  fragments,
  cell_names,
  path,
  insertion_mode = c("both", "start_only", "end_only"),
  threads = 1
)
}
\arguments{
\item{fragments}{IterableFragments object}

\item{cell_names}{Character vector of cluster assignments for each cell. If is null, all cells are treated as one group.}

\item{path}{Path to save MACS2/3 output files. If \code{cell_names} is provided, this must be a character vector with one name for each level in \code{cell_names}}

\item{insertion_mode}{Which fragment ends to use for insertion counts calculation. One of "both", "start_only", or "end_only"}

\item{threads}{Number of threads to use}
}
\description{
Call MACS2/3 peaks using fragments. UNFINISHED.
}
