% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_pipelines.R
\name{project(PipelineBase,IterableMatrix)}
\alias{project(PipelineBase,IterableMatrix)}
\title{Project input data using a fitted pipeline}
\usage{
project(object, x, ...)
}
\arguments{
\item{object}{(PipelineBase) A fitted pipeline object}

\item{x}{(IterableMatrix) Input data to be transformed}
}
\value{
Data projected by the pipeline
}
\description{
Project input data using a fitted pipeline
}
