% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell_utils.R
\name{project}
\alias{project}
\title{Perform a dimensionality reduction on a matrix using a pre-fit DimReduction object.}
\usage{
project(x, mat, ...)
}
\arguments{
\item{x}{DimReduction object.}

\item{mat}{IterableMatrix object.}
}
\value{
IterableMatrix object of the projected data.
}
\description{
Perform a dimensionality reduction on a matrix using a pre-fit DimReduction object.
}
\details{
DimReduction subclasses should use this to project new data with the same features, to project into the same latent space.
All required information to run a projection should be held in x$fitted_params, including pertinent parameters when constructing the DimReduction subclass object.
If there are no rownames, assume that the matrix is in the same order as the original matrix, assuming that they have the same number of features.
If there are rownames, reorder the matrix to match the order of the original matrix
}
