% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{rank_transform}
\alias{rank_transform}
\title{Rank-transform a matrix}
\usage{
rank_transform(mat, axis)
}
\arguments{
\item{mat}{Data matrix (IterableMatrix)}

\item{axis}{Axis to rank values within. "col" to rank values within each column,
and "row" to rank values within each row.}
}
\description{
Rank the values within each row/col of a matrix, and output
the rank values as a new matrix. Rank values are offset such
that the rank of a 0 value is 0, and ties are handled by
averaging ranks.
}
\details{
Note that efficient rank calculation depends on the storage order
of a matrix, so it may be necessary to call transpose_storage_order()
}
\keyword{internal}
