% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_bed}
\alias{read_bed}
\title{Read a bed file file into a data frame}
\usage{
read_bed(
  path,
  additional_columns = character(0),
  backup_url = NULL,
  timeout = 300
)
}
\arguments{
\item{path}{Path to file (or desired save location if backup_url is used)}

\item{additional_columns}{Names for additional columns in the bed file}

\item{backup_url}{If path does not exist, provides a URL to download the gtf from}

\item{timeout}{Maximum time in seconds to wait for download from backup_url}
}
\value{
data frame with coordinates using the 0-based convention
}
\description{
Read a bed file file into a data frame
}
