% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_encode_blacklist}
\alias{read_encode_blacklist}
\title{Download ENCODE blacklist annotations}
\usage{
read_encode_blacklist(
  dir,
  genome = c("hg38", "mm10", "hg19", "dm6", "dm3", "ce11", "ce10"),
  timeout = 300
)
}
\arguments{
\item{dir}{output directory to cache the downloaded file}

\item{genome}{genome name}

\item{timeout}{Maximum time in seconds to wait for download from backup_url}
}
\value{
Data frame with coordinates using the 0-based convention
}
\description{
Downloads the Boyle Lab blacklist, as described in https://doi.org/10.1038/s41598-019-45839-z
}
