% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_gencode_genes}
\alias{read_gencode_genes}
\title{Download gene annotations from gencode}
\usage{
read_gencode_genes(
  dir,
  release = "41",
  gene_type = "lncRNA|protein_coding|IG_.*_gene|TR_.*_gene",
  timeout = 300
)
}
\arguments{
\item{dir}{output directory to cache the downloaded file}

\item{release}{release version (prefix with M for mouse versions). For
most recent version, use "latest" or "latest_mouse"}

\item{gene_type}{Regular expression with which gene types to keep.
Defaults to protein_coding, lncRNA, and IG/TR genes}

\item{timeout}{Maximum time in seconds to wait for download from backup_url}
}
\value{
Data frame with coordinates using the 0-based convention
}
\description{
Download gene annotations from gencode
}
