% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_gtf}
\alias{read_gtf}
\title{Read a gtf file into a data frame}
\usage{
read_gtf(
  path,
  attributes = c("gene_id"),
  features = c("gene"),
  keep_attribute_column = FALSE,
  backup_url = NULL,
  timeout = 300
)
}
\arguments{
\item{path}{Path to file (or desired save location if backup_url is used)}

\item{attributes}{Vector of GTF attribute names to parse out as columns}

\item{features}{List of features types to keep from the GTF (e.g. gene, transcript, exon, intron)}

\item{keep_attribute_column}{Boolean for whether to preserve the raw attribute text column}

\item{backup_url}{If path does not exist, provides a URL to download the gtf from}

\item{timeout}{Maximum time in seconds to wait for download from backup_url}
}
\value{
data frame with coordinates using the 0-based convention
}
\description{
Read a gtf file into a data frame
}
