% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplots.R
\name{render_plot_from_storage}
\alias{render_plot_from_storage}
\title{Render a plot with intermediate disk storage step}
\usage{
render_plot_from_storage(plot, width, height)
}
\arguments{
\item{plot}{(ggplot) ggplot output from a plotting function}

\item{width}{(numeric) width of rendered plot}

\item{height}{(numeric) height of rendered plot}
}
\description{
Take a plotting object and save in temp storage, so it can be outputted with exact dimensions.
Primarily used to allow for adjusting plot dimensions within function reference examples.
}
\keyword{internal}
