% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rotate_x_labels}
\alias{rotate_x_labels}
\title{Rotate ggplot x axis labels}
\usage{
rotate_x_labels(degrees = 45)
}
\arguments{
\item{degrees}{Number of degrees to rotate by}
}
\description{
Rotate ggplot x axis labels
}
\examples{
mat <- get_demo_mat()
cell_types <- paste("Group", rep(1:3, length.out = length(colnames(mat))))

## Plot dot
scale_next_plot_height(0.8)
plot_dot(mat, c("MS4A1", "CD3E"), cell_types)

scale_next_plot_height(0.8)
plot_dot(mat, c("MS4A1", "CD3E"), cell_types) + rotate_x_labels(90)
}
