% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{scan_fragments}
\alias{scan_fragments}
\title{Scan through fragments without performing any operations (used for benchmarking)}
\usage{
scan_fragments(fragments)
}
\arguments{
\item{fragments}{Fragments object to scan}
}
\value{
Length 4 vector with fragment count, then sums of chr, starts, and ends
}
\description{
Scan through fragments without performing any operations (used for benchmarking)
}
\keyword{internal}
