% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{select_cells}
\alias{select_cells}
\title{Select cells for subsetting or translating cells IDs in a fragments object}
\usage{
select_cells(fragments, cell_selection)
}
\arguments{
\item{fragments}{Input fragments object}

\item{cell_selection}{List of chromosme IDs (numeric), or names (character).
The output chromosome ID n will be taken
from the input fragments chromosome with ID/name cell_selection\link{n}.}
}
\description{
Select cells for subsetting or translating cells IDs in a fragments object
}
