% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{select_chromosomes}
\alias{select_chromosomes}
\title{Subset, translate, or reorder chromosome IDs}
\usage{
select_chromosomes(fragments, chromosome_selection)
}
\arguments{
\item{fragments}{Input fragments object}

\item{chromosome_selection}{List of chromosme IDs (numeric), or names (character), or logical mask.}
}
\description{
Subset, translate, or reorder chromosome IDs
}
\details{
Numeric chromosome IDs will be re-assigned in the order of \code{chromosome_selection}.
The output chromosome ID \code{n} will be taken from the input chromosome with ID/name \code{chromosome_selection[n]}.
}
