% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell_utils.R
\name{select_features_by_binned_dispersion}
\alias{select_features_by_binned_dispersion}
\title{Get the most variable features within a matrix.}
\usage{
select_features_by_binned_dispersion(
  mat,
  num_feats = 25000,
  n_bins = 20,
  threads = 1L
)
}
\arguments{
\item{mat}{(IterableMatrix) dimensions features x cells}

\item{num_feats}{(integer) Number of features to return.  If the number is higher than the number of features in the matrix,
all features will be returned.}

\item{n_bins}{(integer) Number of bins for binning mean gene expression.  Normalizing dispersion is done with respect to each bin,
and if the number of features
within a bin is less than 2, the dispersion is set to 1.}

\item{threads}{(integer) Number of threads to use.}
}
\value{
Return a dataframe with the following columns, sorted descending by bin-normalized dispersion:
\itemize{
\item \code{names}: Feature name.
\item \code{score}: Bin-normalized dispersion of the feature.
\item \code{highly_variable}: Logical vector of whether the feature is highly variable.
}
}
\description{
Get the most variable features within a matrix.
}
\details{
The formula for calculating the most variable features is from the Seurat package (Satjia et al. 2015).

Calculate using the following process:
\enumerate{
\item Calculate the dispersion of each feature (variance / mean)
\item Log normalize dispersion and mean
\item Bin the features by their means, and normalize dispersion within each bin
}
}
