% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell_utils.R
\name{select_features_by_dispersion}
\alias{select_features_by_dispersion}
\title{Get the features with the highest dispersion within a matrix.}
\usage{
select_features_by_dispersion(
  mat,
  num_feats = 25000,
  normalize = NULL,
  threads = 1L
)
}
\arguments{
\item{mat}{(IterableMatrix) dimensions features x cells}

\item{num_feats}{(integer) Number of features to return.  If the number is higher than the number of features in the matrix,
all features will be returned.}

\item{normalize}{(function) Normalize matrix using a given function. If \code{NULL}, no normalization is performed.}

\item{threads}{(integer) Number of threads to use.}
}
\value{
Return a dataframe with the following columns, sorted descending by dispersion:
\itemize{
\item \code{names}: Feature name.
\item \code{score}: Variance of the feature.
\item \code{highly_variable}: Logical vector of whether the feature is highly variable.
}
}
\description{
Get the features with the highest dispersion within a matrix.
}
\details{
Calculate using the following process:
\enumerate{
\item Perform an optional term frequency + log normalization, for each feature.
\item Find the dispersion (variance/mean) of each feature.
\item Find \code{num_feats} features with the highest dispersion.
}
}
