% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell_utils.R
\name{select_features_by_mean}
\alias{select_features_by_mean}
\title{Get the top features from a matrix, based on the mean accessibility of each feature.}
\usage{
select_features_by_mean(mat, num_feats = 25000, threads = 1L)
}
\arguments{
\item{mat}{(IterableMatrix) dimensions features x cells}

\item{num_feats}{Number of features to deem as highly accessible.  If the number is higher than the number of features in the matrix,
all features will be returned.}

\item{threads}{(integer) Number of threads to use.}
}
\value{
Return a dataframe with the following columns, sorted descending by mean accessibility:
\itemize{
\item \code{names}: Feature name.
\item \code{score}: Binarize sum of each feature.
\item \code{highly_variable}: Logical vector of whether the feature is highly accessible by mean accessibility.
}
}
\description{
Get the top features from a matrix, based on the mean accessibility of each feature.
}
\details{
Calculate using the following process:
\enumerate{
\item Get the sum of each binarized feature.
\item Find \code{num_feats} features with the highest accessibility.
}
}
