% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell_utils.R
\name{select_features_by_variance}
\alias{select_features_by_variance}
\title{Get the most variable features within a matrix.}
\usage{
select_features_by_variance(
  mat,
  num_feats = 25000,
  normalize = normalize_log,
  threads = 1L
)
}
\arguments{
\item{mat}{(IterableMatrix) dimensions features x cells}

\item{num_feats}{(integer) Number of features to return.  If the number is higher than the number of features in the matrix,
all features will be returned.}

\item{normalize}{(function) Normalize matrix using a given function. If \code{NULL}, no normalization is performed.}

\item{threads}{(integer) Number of threads to use.}
}
\value{
Return a dataframe with the following columns, sorted descending by variance:
\itemize{
\item \code{names}: Feature name.
\item \code{score}: Variance of the feature.
\item \code{highly_variable}: Logical vector of whether the feature is highly variable.
}
}
\description{
Get the most variable features within a matrix.
}
\details{
Calculate using the following process:
\enumerate{
\item Perform an optional term frequency + log normalization, for each feature.
\item Find \code{num_feats} features with the highest variance across clusters.
}
}
