% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{select_regions}
\alias{select_regions}
\title{Select fragments overlapping (or not overlapping) selected regions}
\usage{
select_regions(
  fragments,
  ranges,
  invert_selection = FALSE,
  zero_based_coords = !is(ranges, "GRanges")
)
}
\arguments{
\item{fragments}{Input fragments object.}

\item{ranges}{GRanges object with the ranges to overlap, or list/data frame with columns chr, start, & end.
Must be sorted in order (chr, end, start) where chromosomes are ordered according to the chromosome names of \code{fragments}.}

\item{invert_selection}{If TRUE, select fragments \emph{not} overlapping selected regions
instead of only fragments overlapping the selected regions.}

\item{zero_based_coords}{Whether to convert the ranges from a 1-based end-inclusive
coordinate system to a 0-based end-exclusive coordinate system. Defaults to true
for GRanges and false for other formats
(see \url{http://genome.ucsc.edu/blog/the-ucsc-genome-browser-coordinate-counting-systems/})}
}
\value{
Fragments object filtered according to the selected regions
}
\description{
Select fragments overlapping (or not overlapping) selected regions
}
