% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{selectRegions}
\alias{selectRegions}
\title{Select fragments overlapping (or not overlapping) selected regions}
\usage{
selectRegions(
  fragments,
  ranges,
  invert_selection = FALSE,
  zero_based_coords = TRUE
)
}
\arguments{
\item{fragments}{Input fragments object.}

\item{ranges}{GRanges object with the ranges to overlap, or list/data frame with columns chr, start, & end.
Must be sorted in order (chr, end, start) where chromosomes are ordered according to the chromosome names of \code{fragments}.}

\item{invert_selection}{If TRUE, select fragments \emph{not} overlapping selected regions
instead of only fragments overlapping the selected regions.}

\item{zero_based_coords}{Boolean for whether the input ranges are in a 0-based
or a 1-based coordinate system. (1-based will be converted to 0-based)
(see http://genome.ucsc.edu/blog/the-ucsc-genome-browser-coordinate-counting-systems/)}
}
\value{
Fragments object filtered according to the selected regions
}
\description{
Select fragments overlapping (or not overlapping) selected regions
}
