% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{set_threads}
\alias{set_threads}
\title{Set matrix op thread count}
\usage{
set_threads(mat, threads = 0L)
}
\arguments{
\item{mat}{IterableMatrix, product of rbind or cbind}

\item{threads}{Number of threads to use for execution}
}
\description{
Set number of threads to use for sparse-dense multiply and matrix_stats.
}
\details{
Only valid for concatenated matrices
}
\keyword{internal}
