% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{shift_fragments}
\alias{shift_fragments}
\title{Shift start or end coordinates of fragments by a fixed amount}
\usage{
shift_fragments(fragments, shift_start = 0L, shift_end = 0L)
}
\arguments{
\item{fragments}{Input fragments object}

\item{shift_start}{How many basepairs to shift the start coords}

\item{shift_end}{How many basepairs to shift the end coords}
}
\value{
Shifted fragments object
}
\description{
Shift start or end coordinates of fragments by a fixed amount
}
