% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{svds-internal}
\alias{svds-internal}
\alias{svds.default}
\alias{svds.IterableMatrix}
\title{Internal svds methods}
\usage{
\method{svds}{default}(A, k, nu = k, nv = k, opts = list(), ...)

\method{svds}{IterableMatrix}(A, k, nu = k, nv = k, opts = list(), threads = 0, ...)
}
\description{
Internal svds methods
}
\keyword{internal}
