% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{tile_ranges}
\alias{tile_ranges}
\title{Get ranges corresponding to selected tiles of a tile matrix}
\usage{
tile_ranges(tile_matrix, selection)
}
\description{
Get ranges corresponding to selected tiles of a tile matrix
}
\keyword{internal}
