% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplots.R
\name{trackplot_calculate_segment_height}
\alias{trackplot_calculate_segment_height}
\title{Calculate y positions for transcript segments to avoid overlap
Steps:
\enumerate{
\item Calculate the maximum overlap depth of transcripts
\item Iterate through transcript start/end in sorted order
\item Randomly assign each transcript a y-coordinate between 1 and max overlap depth,
with the restriction that a transcript can't have the same y-coordinate
If is a transcript, filter out the exons in determining overlap so there aren't repeated steps
}}
\usage{
trackplot_calculate_segment_height(data, is_transcript = FALSE)
}
\description{
Calculate y positions for transcript segments to avoid overlap
Steps:
\enumerate{
\item Calculate the maximum overlap depth of transcripts
\item Iterate through transcript start/end in sorted order
\item Randomly assign each transcript a y-coordinate between 1 and max overlap depth,
with the restriction that a transcript can't have the same y-coordinate
If is a transcript, filter out the exons in determining overlap so there aren't repeated steps
}
}
