% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplots.R
\name{trackplot_coverage}
\alias{trackplot_coverage}
\title{Pseudobulk coverage trackplot}
\usage{
trackplot_coverage(
  fragments,
  region,
  groups,
  cell_read_counts,
  group_order = NULL,
  bins = 500,
  clip_quantile = 0.999,
  colors = discrete_palette("stallion"),
  legend_label = NULL,
  zero_based_coords = !is(region, "GRanges"),
  return_data = FALSE
)
}
\arguments{
\item{fragments}{Fragments object}

\item{region}{Region to plot, e.g. output from \code{gene_region()}. String of format "chr1:100-200",
or list/data.frame/GRanges of length 1 specifying chr, start, end. See \code{help("genomic-ranges-like")} for details}

\item{groups}{Vector with one entry per cell, specifying the cell's group}

\item{cell_read_counts}{Numeric vector of read counts for each cell (used for normalization)}

\item{group_order}{Optional vector listing ordering of groups}

\item{bins}{Number of bins to plot across the region}

\item{clip_quantile}{(optional) Quantile of values for clipping y-axis limits. Default of 0.999 will crop out
just the most extreme outliers across the region. NULL to disable clipping}

\item{colors}{Character vector of color values (optionally named by group)}

\item{legend_label}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Custom label to put on the legend (no longer used as color legend is not shown anymore)}

\item{zero_based_coords}{Whether to convert the ranges from a 1-based end-inclusive
coordinate system to a 0-based end-exclusive coordinate system. Defaults to true
for GRanges and false for other formats
(see this \href{https://web.archive.org/web/20210920203703/https://genome.ucsc.edu/blog/the-ucsc-genome-browser-coordinate-counting-systems/}{archived UCSC blogpost})}

\item{return_data}{If true, return data from just before plotting rather than a plot.}
}
\value{
Returns a combined plot of pseudobulk genome tracks. For compatability with
\code{draw_trackplot_grid()}, the extra attribute \verb{$patches$labels} will be added to
specify the labels for each track. If \code{return_data} or \code{return_plot_list} is
\code{TRUE}, the return value will be modified accordingly.
}
\description{
Plot a pseudobulk genome track, showing the number of fragment insertions
across a region for each cell type or group.
}
\examples{
## Prep data
frags <- get_demo_frags()
cell_types <- paste("Group", rep(1:3, length.out = length(cellNames(frags))))

\dontrun{
## Use genes and blacklist to determine proper number of reads per cell
genes <- read_gencode_transcripts(
  file.path(tempdir(), "references"), release = "42",
  annotation_set = "basic",
  features = "transcript", timeout = 3000
)
blacklist <- read_encode_blacklist(file.path(tempdir(), "references"), genome="hg38")

read_counts <- qc_scATAC(frags, genes, blacklist)$nFrags
}
region <- "chr4:3034877-4034877"

## We use pre-generated data for this example
read_counts <- readr::read_delim(
 file.path(system.file("extdata", package = "BPCells"), 
  "qc_results_filtered_example_chr_4.tsv.gz"),
 delim = "\t", show_col_types = FALSE
)$nFrags


scale_next_plot_height(0.5)
trackplot_coverage(
  frags, region, groups = cell_types, 
  cell_read_counts = read_counts
)
}
\seealso{
\code{trackplot_combine()}, \code{trackplot_gene()}, \code{trackplot_loop()}, \code{trackplot_scalebar()}
}
