% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplots.R
\name{trackplot_create_arrow_segs}
\alias{trackplot_create_arrow_segs}
\title{Break up segments into smaller segments the length of the plot, divided by size}
\usage{
trackplot_create_arrow_segs(data, region, size = 150, extra_cols = c())
}
\arguments{
\item{data}{Dataframe of full segments to be broken up}

\item{region}{Region to be plotted with end and start attr}

\item{size}{int Number of arrows to span the x axis of track}

\item{extra_cols}{Vector of characters of columns to keep from data}
}
\description{
Break up segments into smaller segments the length of the plot, divided by size
}
