% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplots.R
\name{trackplot_gene}
\alias{trackplot_gene}
\title{Plot transcript models}
\usage{
trackplot_gene(
  transcripts,
  region,
  exon_size = 2.5,
  gene_size = 0.5,
  label_size = 11 * 0.8/ggplot2::.pt,
  track_label = "Genes",
  return_data = FALSE
)
}
\arguments{
\item{transcripts}{Transcipt features given as GRanges, data.frame, or list. See \code{help("genomic-ranges-like")} for details on format and coordinate systems. Required attributes:
\itemize{
\item \code{chr}, \code{start}, \code{end}: genomic position
\item \code{strand}: +/- or TRUE/FALSE for positive or negative strand
\item \code{feature}: Only entries marked as \code{"transcript"} or \code{"exon"} will be considered
\item \code{gene_name}: Symbol or gene ID to display
\item \code{transcript_id}: Transcritp identifier to link transcripts and exons
}

Usually given as the output from \code{read_gencode_transcripts()}}

\item{region}{Region to plot, e.g. output from \code{gene_region()}. String of format "chr1:100-200",
or list/data.frame/GRanges of length 1 specifying chr, start, end. See \code{help("genomic-ranges-like")} for details}

\item{exon_size}{size for exon lines in units of mm}

\item{gene_size}{size for intron/gene lines in units of mm}

\item{label_size}{size for transcript labels in units of mm}

\item{track_label}{Label to put on the side of the track}

\item{return_data}{If true, return data from just before plotting rather than a plot.}
}
\value{
Plot of gene locations
}
\description{
Plot transcript models
}
\examples{
\dontrun{
## Prep data
transcripts <- read_gencode_transcripts(
  file.path(tempdir(), "references"), release = "42",
  annotation_set = "basic", features = "transcript", timeout = 3000
)
}

## We use pre-generated data for this example
transcripts <- readr::read_delim(
  file.path(system.file("extdata", package = "BPCells"),
  "transcripts_filtered_example_chr_4.tsv.gz"))

region <- "chr4:3264877-3634877" 

## Plot gene trackplot
scale_next_plot_height(0.3)
trackplot_gene(transcripts, region)
}
\seealso{
\code{trackplot_combine()}, \code{trackplot_coverage()}, \code{trackplot_loop()}, \code{trackplot_scalebar()}
}
