% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplots.R
\name{trackplot_genome_annotation}
\alias{trackplot_genome_annotation}
\title{Plot range-based annotation tracks (e.g. peaks)}
\usage{
trackplot_genome_annotation(
  loci,
  region,
  color_by = NULL,
  colors = NULL,
  label_by = NULL,
  label_size = 11 * 0.8/ggplot2::.pt,
  show_strand = FALSE,
  annotation_size = 3,
  track_label = "Peaks",
  return_data = FALSE
)
}
\arguments{
\item{loci}{Genomic loci given as GRanges, data.frame, or list. See \code{help("genomic-ranges-like")} for details on format and coordinate systems. Required attributes:
\itemize{
\item \code{chr}, \code{start}, \code{end}: genomic position
}}

\item{region}{Region to plot, e.g. output from \code{gene_region()}. String of format "chr1:100-200",
or list/data.frame/GRanges of length 1 specifying chr, start, end. See \code{help("genomic-ranges-like")} for details}

\item{color_by}{Name of a metadata column in \code{loci} to use for coloring, or a data vector with same length as loci. Column must be numeric or convertible to a factor.}

\item{colors}{Vector of hex color codes to use for the color scale. For numeric \code{color_by} data, this is passed to \code{ggplot2::scale_color_gradientn()},
otherwise it is interpreted as a discrete color palette in \code{ggplot2::scale_color_manual()}}

\item{label_by}{Name of a metadata column in \code{loci} to use for labeling, or a data vector with same length as loci. Column must hold string data.}

\item{label_size}{size for labels in units of mm}

\item{show_strand}{If TRUE, show strand direction as arrows}

\item{annotation_size}{size for annotation lines in mm}

\item{return_data}{If true, return data from just before plotting rather than a plot.}
}
\value{
Plot of genomic loci if return_data is FALSE, otherwise returns the data frame used to generate the plot
}
\description{
Plot range-based annotation tracks (e.g. peaks)
}
\seealso{
\code{trackplot_combine()}, \code{trackplot_coverage()}, \code{trackplot_loop()}, \code{trackplot_scalebar()}, \code{trackplot_gene()}
}
