% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplots.R
\name{trackplot_genome_annotations}
\alias{trackplot_genome_annotations}
\title{Plot peak tracks}
\usage{
trackplot_genome_annotations(
  loci,
  region,
  annotation_size = 4,
  label_by = NULL,
  label_size = 11 * 0.8/ggplot2::.pt,
  color_by = NULL,
  colors = NULL,
  direction_by = NULL,
  arrow_angle = 75,
  track_label = "Peaks",
  return_data = FALSE
)
}
\arguments{
\item{loci}{Genomic loci given as GRanges, data.frame, or list. See \code{help("genomic-ranges-like")} for details on format and coordinate systems. Required attributes:
\itemize{
\item \code{chr}, \code{start}, \code{end}: genomic position
}}

\item{region}{Region to plot, e.g. output from \code{gene_region()}. String of format "chr1:100-200",
or list/data.frame/GRanges of length 1 specifying chr, start, end. See \code{help("genomic-ranges-like")} for details}

\item{annotation_size}{size for peak annotations}

\item{label_by}{Name of a metadata column in \code{loci} to use for labeling, or a numeric vector with same length as loci  Column must be type character, factor or numeric}

\item{label_size}{size for transcript labels in units of mm}

\item{color_by}{Name of a metadata column in \code{loci} to use for coloring, or a numeric vector with same length as loci.  Column must be type character, factor, or numeric}

\item{colors}{Vector of hex color codes to use for the color gradient}

\item{direction_by}{Name of a metadata column in \code{loci} to use for determining direction of arrows}

\item{arrow_angle}{Angle of the arrow head in degrees, if direction_by is provided}

\item{track_label}{Label for the track}

\item{return_data}{If TRUE, return the data frame used to generate the plot}
}
\value{
Plot of genomic loci if return_data is FALSE, otherwise returns the data frame used to generate the plot
}
\description{
Plot peak tracks
}
\seealso{
\code{trackplot_combine()}, \code{trackplot_coverage()}, \code{trackplot_loop()}, \code{trackplot_scalebar()}, \code{trackplot_gene()}
}
