% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplots.R
\name{trackplot_normalize_ranges_with_metadata}
\alias{trackplot_normalize_ranges_with_metadata}
\title{Normalize trackplot ranges data, while handling metadata argument renaming and type conversions
Type conversions are as follows:
color -> factor or numeric
label -> string}
\usage{
trackplot_normalize_ranges_with_metadata(data, metadata)
}
\arguments{
\item{data}{Input ranges-like object}

\item{metadata}{List of form e.g. list(color=color_by, label=label_by). The values can be either column names or data vectors.
Any NULL values will be skipped}
}
\value{
Tibble with normalized ranges and additional columns populated as requested in \code{metadata}
}
\description{
Normalize trackplot ranges data, while handling metadata argument renaming and type conversions
Type conversions are as follows:
color -> factor or numeric
label -> string
}
\keyword{internal}
