% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplots.R
\name{trackplot_scalebar}
\alias{trackplot_scalebar}
\title{Plot scale bar}
\usage{
trackplot_scalebar(region, font_pt = 11)
}
\arguments{
\item{region}{Region to plot, e.g. output from \code{gene_region()}. String of format "chr1:100-200",
or list/data.frame/GRanges of length 1 specifying chr, start, end. See \code{help("genomic-ranges-like")} for details}

\item{font_pt}{Font size for scale bar labels in units of pt.}
}
\value{
Plot with coordinates and scalebar for plotted genomic region
}
\description{
Plots a human-readable scale bar and coordinates of the region being plotted
}
\examples{
region <- "chr4:3034877-3044877"
BPCells:::render_plot_from_storage(
  trackplot_scalebar(region), width = 6, height = 1
)
}
\seealso{
\code{trackplot_combine()}, \code{trackplot_coverage()}, \code{trackplot_gene()}, \code{trackplot_loop()}
}
