% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_project_predict.R
\name{transform(PipelineBase,IterableMatrix)}
\alias{transform(PipelineBase,IterableMatrix)}
\title{Transform the input data using a fitted pipeline}
\usage{
transform(object, x, ...)
}
\arguments{
\item{object}{(PipelineBase) The fitted pipeline object}

\item{x}{(IterableMatrix) Input data to be transformed}
}
\value{
Data transformed by the pipeline
}
\description{
Transform the input data using a fitted pipeline
}
