% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_transformers.R
\docType{class}
\name{VarFeatSelectorTransformer}
\alias{VarFeatSelectorTransformer}
\title{Perform feature selection on a matrix using dispersion.
TODO: Add more details from when upstream lsi PR is reviewed.}
\usage{
VarFeatSelectorTransformer(num_feats, n_bins = 20L, threads = 1L)
}
\arguments{
\item{num_feats}{(integer) Number of features to return.  If the number is higher than the number of features in the matrix,
all features will be returned.}

\item{n_bins}{(integer) Number of bins for binning mean gene expression.  Normalizing dispersion is done with respect to each bin,
and if the number of features
within a bin is less than 2, the dispersion is set to 1.}

\item{threads}{(integer) Number of threads to use.}
}
\description{
Perform feature selection on a matrix using dispersion.
TODO: Add more details from when upstream lsi PR is reviewed.

Get the most variable features within a matrix.
}
\details{
The formula for calculating the most variable features is from the Seurat package (Satjia et al. 2015).

Calculate using the following process:
\enumerate{
\item Calculate the dispersion of each feature (variance / mean)
\item Log normalize dispersion and mean
\item Bin the features by their means, and normalize dispersion within each bin
}
}
