% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{wrapMatrix}
\alias{wrapMatrix}
\title{Construct an S4 matrix object wrapping another matrix object, avoiding
duplicate storage of dimnames}
\usage{
wrapMatrix(class, m, ...)
}
\description{
Construct an S4 matrix object wrapping another matrix object, avoiding
duplicate storage of dimnames
}
