% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{write_fragments_10x}
\alias{write_fragments_10x}
\title{Write to a 10x fragments file}
\usage{
write_fragments_10x(
  fragments,
  path,
  end_inclusive = TRUE,
  append_5th_column = FALSE
)
}
\arguments{
\item{fragments}{Input fragments object}

\item{path}{Path of 10x fragments file}

\item{end_inclusive}{Whether the end coordinate of the bed is inclusive -- i.e. there was an
insertion at the end coordinate rather than the base before the end coordinate. This is the
10x default, though it's not quite standard for the bed file format.}

\item{append_5th_column}{Whether to include 5th column of all 0 for compatibility
with 10x fragment file outputs (defaults to 4 columns chr,start,end,cell)}
}
\description{
Write to a 10x fragments file
}
