% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{write_fragments_dir}
\alias{write_fragments_dir}
\title{Make a fragments object in binary files within a directory.}
\usage{
write_fragments_dir(fragments, dir, compress = TRUE, buffer_size = 1024L)
}
\arguments{
\item{fragments}{Input fragments object}

\item{dir}{Directory to save the data into}

\item{compress}{Whether or not to compress the data. With compression, memory usage should
be about half the size of a gzip-compressed 10x fragments file holding the same data.
Without compression memory usage can be 10-20\% larger.}

\item{buffer_size}{For performance tuning only. The number of items to be bufferred
in memory before calling writes to disk.}
}
\value{
FragmentsDir object
}
\description{
Make a fragments object in binary files within a directory.
}
