% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{write_fragments_hdf5}
\alias{write_fragments_hdf5}
\title{Write a fragments object in an hdf5 file}
\usage{
write_fragments_hdf5(
  fragments,
  path,
  group = "fragments",
  compress = TRUE,
  buffer_size = 8192L,
  chunk_size = 1024L
)
}
\arguments{
\item{fragments}{Input fragments object}

\item{path}{Path to the hdf5 file on disk}

\item{group}{The group within the hdf5 file to write the data to. If writing
to an existing hdf5 file this group must not already be in use}

\item{compress}{Whether or not to compress the data. With compression, memory usage should
be about half the size of a gzip-compressed 10x fragments file holding the same data.
Without compression memory usage can be 10-20\% larger.}

\item{buffer_size}{For performance tuning only. The number of items to be bufferred
in memory before calling writes to disk.}

\item{chunk_size}{For performance tuning only. The chunk size used for the HDF5 array storage.}
}
\value{
FragmentsHDF5 object
}
\description{
Write a fragments object in an hdf5 file
}
