% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{write_fragments_memory}
\alias{write_fragments_memory}
\title{Make a fragments object in memory.}
\usage{
write_fragments_memory(fragments, compress = TRUE)
}
\arguments{
\item{fragments}{Input fragments object}

\item{compress}{Whether or not to compress the data. With compression, memory usage should
be about half the size of a gzip-compressed 10x fragments file holding the same data.
Without compression memory usage can be 10-20\% larger.}
}
\value{
MemFragments object
}
\description{
Make a fragments object in memory.
}
