% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac_utils.R
\name{write_insertion_bed}
\alias{write_insertion_bed}
\title{Create bed files from fragments split by cell group.}
\usage{
write_insertion_bed(
  fragments,
  path,
  cell_groups = rlang::rep_along(cellNames(fragments), "all"),
  insertion_mode = c("start_only", "both", "end_only"),
  verbose = FALSE,
  threads = 1
)
}
\arguments{
\item{fragments}{IterableFragments object}

\item{path}{(character vector) Path to save bed files. If \code{cell_groups} is provided, this must be a character vector with one name for each level in \code{cell_groups}
Else, this must be a character vector of length 1.}

\item{cell_groups}{(character vector or factor) Cluster assignments for each cell.}

\item{insertion_mode}{(string) Which fragment ends to use for coverage calculation. One of "both", "start_only", or "end_only"}

\item{verbose}{(bool) Whether to provide verbose progress output to console.}

\item{threads}{(int) Number of threads to use.}
}
\value{
\code{NULL}
}
\description{
Create bed files from fragments split by cell group.
}
\keyword{internal}
