% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{write_matrix_10x_hdf5}
\alias{write_matrix_10x_hdf5}
\title{Write a matrix to a 10x-compatible hdf5 feature matrix file}
\usage{
write_matrix_10x_hdf5(
  mat,
  path,
  barcodes = colnames(mat),
  feature_ids = rownames(mat),
  feature_names = rownames(mat),
  feature_types = "Gene Expression",
  feature_metadata = list(),
  buffer_size = 16384L,
  chunk_size = 1024L
)
}
\arguments{
\item{mat}{IterableMatrix}

\item{path}{Path to the hdf5 file on disk}

\item{barcodes}{Vector of names for the cells}

\item{feature_ids}{Vector of IDs for the features}

\item{feature_names}{Vector of names for the features}

\item{feature_metadata}{Named list of additional metadata vectors
to store for each feature}

\item{buffer_size}{For performance tuning only. The number of items to be buffered
in memory before calling writes to disk.}

\item{chunk_size}{For performance tuning only. The chunk size used for the HDF5 array storage.}

\item{feature_type}{String or vector of feature types}
}
\value{
IterableMatrix with the contents of the output
}
\description{
Write a matrix to a 10x-compatible hdf5 feature matrix file
}
\details{
Input matrices must be in column-major storage order,
and if the rownames and colnames are not set, names must be
provided for the relevant metadata parameters. Some of the
metadata parameters are not read by default in BPCells, but
it is possible to export them for use with other tools.
}
