% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\docType{class}
\name{XPtrList-class}
\alias{XPtrList-class}
\title{This class is used to hold a list of external pointers.}
\description{
This class is used to hold a list of external pointers.
}
\keyword{internal}
