% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{all_matrix_inputs}
\alias{all_matrix_inputs}
\alias{all_matrix_inputs<-}
\title{Get/set inputs to a matrix transform}
\usage{
all_matrix_inputs(x)

all_matrix_inputs(x) <- value
}
\arguments{
\item{x}{IterableMatrix}

\item{value}{List of IterableMatrix objects}
}
\value{
List of IterableMatrix objects. If a matrix \code{m} is itself an input object, then
\code{all_matrix_inputs(m)} will return \code{list(m)}.
}
\description{
A matrix object can either be an input (i.e. a file on disk or a raw matrix in memory),
or it can represent a delayed operation on one or more matrices. The \code{all_matrix_inputs()}
getter and setter functions allow accessing the base-level input matrices as a list, and
changing them. This is useful if you want to re-locate data on disk without losing your
transformed BPCells matrix. (Note: experimental API; potentially subject to revisions).
}
