% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{binarize}
\alias{binarize}
\title{Binarize converts matrix elements to zeros and ones.}
\usage{
binarize(x, threshold = 1)
}
\arguments{
\item{x}{matrix An input matrix.}

\item{threshold}{A numeric value that determines whether the elements
of x are set to zero or one.}
}
\value{
matrix
}
\description{
Binarize compares the matrix element values to the
threshold value. Elements whose value is less than the threshold
are set to zero and values greater than or equal to the
threshold are set to one. The default threshold is 1.
}
