% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac_utils.R
\name{call_macs_peaks}
\alias{call_macs_peaks}
\title{Call MACS2/3 peaks using fragments. UNFINISHED.}
\usage{
call_macs_peaks(
  fragments,
  cell_names,
  genome_size,
  path,
  insertion_mode = c("both", "start_only", "end_only"),
  step = c("prep-inputs", "run-macs", "read-outputs"),
  macs_version = c("macs2", "macs3"),
  threads = 1
)
}
\arguments{
\item{fragments}{IterableFragments object}

\item{cell_names}{Character vector of cluster assignments for each cell. If is null, all cells are treated as one group.}

\item{genome_size}{Effective genome size for MACS2/3}

\item{path}{Directory to save MACS output files.  Stores all clusters in one directory, with the name of the cluster as the prefix.}

\item{insertion_mode}{Which fragment ends to use for insertion counts calculation. One of "both", "start_only", or "end_only"}

\item{step}{Which step to run. One of  "prep-inputs", "run-macs", "read-outputs".  If "prep-inputs", create the input bedfiles for macs,
and provides a shell script per cluster with the command to run macs.  If run-macs, also run bash scripts to execute macs.
If read-outputs, read the outputs into tibbles.}

\item{macs_version}{Which version of MACS to use. One of "macs2" or "macs3"}

\item{threads}{Number of threads to use}
}
\description{
Call MACS2/3 peaks using fragments. UNFINISHED.
}
