% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_graph_leiden}
\alias{cluster_graph_leiden}
\alias{cluster_graph_louvain}
\alias{cluster_graph_seurat}
\title{Cluster an adjacency matrix}
\usage{
cluster_graph_leiden(snn, resolution = 0.001, seed = 12531, ...)

cluster_graph_louvain(snn, resolution = 1, seed = 12531)

cluster_graph_seurat(snn, resolution = 0.8, ...)
}
\arguments{
\item{snn}{Symmetric adjacency matrix (dgCMatrix) output from e.g. knn_to_snn_graph. Only the lower triangle is used}

\item{resolution}{Resolution parameter. Higher values result in more clusters}

\item{seed}{Random seed for clustering initialization}

\item{...}{Additional arguments to underlying clustering function}
}
\value{
Factor vector containing the cluster assignment for each cell.
}
\description{
Cluster an adjacency matrix
}
\details{
\strong{cluster_graph_leiden}: Leiden graph clustering algorithm \code{igraph::cluster_leiden()}

\strong{cluster_graph_louvain}: Louvain graph clustering algorithm \code{igraph::cluster_louvain()}

\strong{cluster_graph_seurat}: Seurat's clustering algorithm \code{Seurat::FindClusters()}
}
