% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_cells_graph}
\alias{cluster_cells_graph}
\title{Cluster embeddings using a KNN-Graph based algorithm}
\usage{
cluster_cells_graph(
  mat,
  knn_method = knn_hnsw,
  knn_to_graph_method = knn_to_geodesic_graph,
  graph_to_cluster_method = cluster_graph_leiden,
  threads = 0L,
  verbose = FALSE
)
}
\arguments{
\item{mat}{(matrix) Cell embeddings matrix of shape \verb{(cells x n_embeddings)}}

\item{knn_method}{(function) Function to convert cell embeddings into a knn object.
Must be a (optionally partialized) version of \code{knn_hnsw()} or \code{knn_annoy()}.}

\item{knn_to_graph_method}{(function) Function to convert the knn object returned from \code{knn_method} to a graph adjacency matrix.
Must be a (optionally partialized) version of \code{knn_to_graph()}, \code{knn_to_snn_graph()} or \code{knn_to_geodesic_graph()}.}

\item{graph_to_cluster_method}{(function) Clustering algorithm that converts a graph adjacency matrix
returned from \code{graph_to_cluster_method} into cluster labels for each cell.
Must be a (optionally partialized) version of \code{cluster_graph_leiden()}, \code{cluster_graph_louvain()} or \code{cluster_graph_seurat()}.}

\item{threads}{(integer) Number of threads to use in \code{knn_method}, \code{knn_to_graph_method} and \code{graph_to_cluster_method}.  If these functions do not utilize
a \code{threads} argument, this is silently ignored.}

\item{verbose}{(logical) Whether to print progress information in \code{knn_method}, \code{knn_to_graph_method} and \code{graph_to_cluster_method}.  If these functions do not utilize
a \code{verbose} argument, this is silently ignored.}
}
\value{
(factor) Factor vector containing the cluster assignment for each cell.
}
\description{
Take in a cell embedding matrix, and sequentially convert it into a kNN object, then to
a graph adjacency matrix.  Following, assign a label to every cell using a clustering algorithm.
}
\details{
\code{cluster_cells_graph()} acts as a helper function to wrap input creation and \code{kNN} graph adjacency-based clustering to be done together.  The user
can also manually pass cell embeddings to their preferred knn/clustering functions of choices.

\strong{Clustering customization through partialized parameters}

Customization of clustering is possible through partialization of each parameter in \code{cluster_cells_graph()} that is a function.
In detail, each parameter that requests a function
may take in one with only some of the arguments provided.  If the first argument is not provided, a copy of a function is utilized that has its parameters
changed with the arguments provided.

For instance, if the user desires for \code{cluster_cells_graph()} to instead use \code{cluster_graph_louvain()} with resolution different than the default,
they can instead call \code{cluster_cells_graph()} like so:
\code{cluster_cells_graph(mat, graph_to_cluster_method = cluter_graph_louvain(resolution = 0.5))}
}
\seealso{
\code{knn_hnsw()} \code{knn_annoy()} \code{knn_to_graph()} \code{knn_to_snn_graph()} \code{knn_to_geodesic_graph()} \code{cluster_graph_leiden()} \code{knn_to_snn_graph()} \code{knn_to_geodesic_graph()}
}
