% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{collect_features}
\alias{collect_features}
\title{Collect features for plotting}
\usage{
collect_features(
  source,
  features = NULL,
  gene_mapping = human_gene_mapping,
  n = 1
)
}
\arguments{
\item{source}{Matrix or data frame to pull features from, or a vector of feature values for a single feature. For
a matrix, the features must be rows.}

\item{features}{Character vector of features names to plot if \code{source} is not a vector.}

\item{gene_mapping}{An optional vector for gene name matching with match_gene_symbol().
Ignored if source is a data frame.}

\item{n}{Internal-use parameter marking the number of nested calls. This is used for
finding the name of the "source" input variable from the caller's perspective}
}
\value{
Data frame with one column for each feature requested
}
\description{
Helper function for data on features to plot from a diverse
set of data sources.
}
\details{
If \code{source} is a data.frame, features will be drawn from the columns. If
\code{source} is a matrix object (\code{IterableMatrix}, \code{dgCMatrix}, or \code{matrix}), features
will be drawn from rows.
}
