% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{convert_mat_to_cluster_matrix}
\alias{convert_mat_to_cluster_matrix}
\title{Convert a matrix to the required input type for clustering}
\usage{
convert_mat_to_cluster_matrix(
  mat,
  required_mat_type = c("knn", "adjacency"),
  knn_mat_method = knn_hnsw,
  knn_graph_method = knn_to_geodesic_graph
)
}
\arguments{
\item{mat}{Input matrix to be converted.}

\item{required_mat_type}{(character) Required matrix type: "adjacency" or "knn".}

\item{knn_mat_method}{(function) Function to convert to a knn object (e.g., \code{knn_hnsw}, \code{knn_annoy}).
Ignored if \code{mat} is already a knn or graph matrix.}

\item{knn_graph_method}{(function) Function to convert a knn object to a graph matrix
(e.g., \code{knn_to_geodesic_graph}). Ignored if \code{required_mat_type} is "knn" or if
\code{mat} is already a graph matrix.}
}
\value{
The converted matrix object.
}
\description{
Ensures that the input matrix is converted to the correct type (knn or adjacency)
required by a clustering function. If the input is already of the correct type,
it is returned as is.
}
\details{
This function checks the type of the input matrix \code{mat}. \code{mat} is returned without modification if
it is already the required type (adjacency or knn).
If \code{mat} is a standard matrix and \code{required_mat_type} is "knn", then \code{knn_mat_method} is used to convert \code{mat} to a knn object.
If \code{required_mat_type} is "adjacency", then \code{knn_mat_method} is used to first convert \code{mat} to a knn object,
then \code{knn_graph_method} is used to convert the knn object to an adjacency matrix.
}
\keyword{internal}
