% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_partial}
\alias{create_partial}
\title{Helper to create partial functions}
\usage{
create_partial()
}
\value{
A \code{bpcells_partial} object (a function with some extra attributes)
}
\description{
Automatically creates a partial application of the caller
function including all non-missing arguments.
}
\keyword{internal}
