% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell_utils.R
\name{DimReduction}
\alias{DimReduction}
\title{Barebones definition of a DimReduction class.}
\usage{
DimReduction(x, fitted_params = list(), ...)
}
\description{
Represents a latent space output of a matrix after a transformation function, with any required information to reproject other inputs using this object.
Child classes should implement a \code{project} method to allow for the projection of other matrices using
the fitted transformation object.
}
\section{Fields}{

\describe{
\item{\code{cell_embeddings}}{(IterableMatrix, dgCMatrix, matrix) Projected data of shape \verb{(n_dimesions x n_cells)} of the original matrix after a dimensionality reduction.}

\item{\code{fitted_params}}{(list) A list of parameters used for the transformation of a matrix.  This should include all necessary information to project new data with the same features.}

\item{\code{feature_names}}{(character) The names of the features that this DimReduction object was fit on.  Matrices to be projected should have the same feature names.}
}}

