% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_project_predict.R
\docType{class}
\name{Estimator}
\alias{Estimator}
\title{S4 Class representing an operation that predicts data, and holds fitted parameters.}
\description{
S4 Class representing an operation that predicts data, and holds fitted parameters.
}
\section{Slots}{

\describe{
\item{\code{step_name}}{(character) Name of the step}

\item{\code{fitted}}{(logical) Whether the pipeline has been fitted}
}}

