% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_project_predict.R
\name{fit(PipelineBase,IterableMatrix)}
\alias{fit(PipelineBase,IterableMatrix)}
\title{Fit the pipeline object to data}
\usage{
fit(object, x, y = NULL, ...)
}
\arguments{
\item{object}{(PipelineBase) The pipeline object to fit.}

\item{x}{(IterableMatrix) Input data to be fitted on.}

\item{y}{Optional output data to be fitted on.  Required if the final step is an Estimator, else ignored.}
}
\value{
The fitted pipeline object.
}
\description{
Fit the pipeline object to data
}
