% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{convert_to_fragments}
\alias{convert_to_fragments}
\title{Convert between BPCells fragments and R objects.}
\usage{
# Convert from R to BPCells
convert_to_fragments(x, zero_based_coords = !is(x, "GRanges"))
as(x, "IterableFragments")

# Convert from BPCells to R
as.data.frame(bpcells_fragments)
as(bpcells_fragments, "data.frame")
as(bpcells_fragments, "GRanges")
}
\arguments{
\item{x}{Fragment coordinates given as GRanges, data.frame, or list. See \code{help("genomic-ranges-like")} for details on format and coordinate systems. Required attributes:
\itemize{
\item \code{chr}, \code{start}, \code{end}: genomic position
\item \code{cell_id}: cell barcodes or unique identifiers as string or factor
}}

\item{zero_based_coords}{Whether to convert the ranges from a 1-based end-inclusive
coordinate system to a 0-based end-exclusive coordinate system. Defaults to true
for GRanges and false for other formats
(see this \href{https://web.archive.org/web/20210920203703/http://genome.ucsc.edu/blog/the-ucsc-genome-browser-coordinate-counting-systems/}{archived UCSC blogpost})}
}
\value{
\strong{convert_to_fragments()}: IterableFragments object
}
\description{
BPCells fragments can be interconverted with GRanges and data.frame R objects.
The main conversion method is R's builtin \code{as()} function, though the
\code{convert_to_fragments()} helper is also available. For all R objects except
GRanges, BPCells assumes a 0-based, end-exclusive coordinate system. (See
\link{genomic-ranges-like} reference for details)
}
