% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{fragments_identical}
\alias{fragments_identical}
\title{Check if two fragments objects are identical}
\usage{
fragments_identical(fragments1, fragments2)
}
\arguments{
\item{fragments1}{First IterableFragments to compare}

\item{fragments2}{Second IterableFragments to compare}
}
\value{
boolean for whether the fragments objects are identical
}
\description{
Check if two fragments objects are identical
}
