% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeAnnotations.R
\name{gene_region}
\alias{gene_region}
\title{Find gene region}
\usage{
gene_region(
  genes,
  gene_symbol,
  extend_bp = c(10000, 10000),
  gene_mapping = human_gene_mapping
)
}
\arguments{
\item{genes}{Transcipt features given as GRanges, data.frame, or list. See \code{help("genomic-ranges-like")} for details on format and coordinate systems. Required attributes:
\itemize{
\item \code{chr}, \code{start}, \code{end}: genomic position
\item \code{strand}: +/- or TRUE/FALSE for positive or negative strand
\item \code{gene_name}: Symbol or gene ID
}}

\item{gene_symbol}{Name of gene symbol or ID}

\item{extend_bp}{Bases to extend region upstream and downstream of gene. If length 1, extension is
symmetric. If length 2, provide upstream extension then downstream extension as positive distances.}

\item{gene_mapping}{Named vector where names are gene symbols or IDs and values are
canonical gene symbols}
}
\value{
List of chr, start, end positions for use with trackplot functions.
}
\description{
Conveniently look up the region of a gene by gene symbol. The value returned by this function
can be used as the \code{region} argument for trackplot functions such as
\code{trackplot_coverage()} or \code{trackplot_gene()}
}
\examples{
## Prep data
genes <- read_gencode_transcripts(
  file.path(tempdir(), "references"), release = "42",
  annotation_set = "basic",
  features = "transcript"
)

## Get gene region
gene_region(genes, "CD19", extend_bp = 1e5)
}
