% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeAnnotations.R
\name{gene_region}
\alias{gene_region}
\title{Find gene region}
\usage{
gene_region(
  genes,
  gene_symbol,
  extend_bp = 10000,
  gene_mapping = human_gene_mapping
)
}
\arguments{
\item{genes}{GRanges, list, or data.frame of transcript features to plot.
Required attributes are:
\itemize{
\item \code{chr}, \code{start}, \code{end}: genomic position
\item \code{gene_name}: Symbol or gene ID
}}

\item{gene_symbol}{Name of gene symbol or ID}

\item{extend_bp}{Bases to extend region upstream and downstream of gene}

\item{gene_mapping}{Named vector where names are gene symbols or IDs and values are
canonical gene symbols}
}
\value{
list of chr, start, end positions for use with , etc.
}
\description{
Conveniently look up the region of a gene by gene symbol. The value returned by this function
can be used as the \code{region} argument for trackplot functions such as
\code{trackplot_bulk()} or \code{trackplot_gene()}
}
