% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneScores.R
\name{gene_score_weights_archr}
\alias{gene_score_weights_archr}
\alias{gene_score_archr}
\title{Calculate GeneActivityScores}
\usage{
gene_score_weights_archr(
  genes,
  chromosome_sizes,
  blacklist = NULL,
  tile_width = 500,
  gene_name_column = "gene_id",
  addArchRBug = FALSE
)

gene_score_archr(
  fragments,
  genes,
  chromosome_sizes,
  blacklist = NULL,
  tile_width = 500,
  gene_name_column = "gene_id",
  addArchRBug = FALSE,
  tile_max_count = 4,
  scale_factor = 10000,
  tile_matrix_path = tempfile(pattern = "gene_score_tile_mat")
)
}
\arguments{
\item{genes}{GRanges object with gene start, end, and strand}

\item{chromosome_sizes}{\link{genomic-ranges} holding start and end coordinates for
each chromosome. See read_ucsc_chrom_sizes().}

\item{blacklist}{Blacklist ranges to exclude from gene calculations. \link{genomic-ranges}
object (list, data.frame, or GRanges)}

\item{tile_width}{Size of tiles to consider}

\item{gene_name_column}{If not NULL, a column name of \code{genes} to use as row names}

\item{addArchRBug}{Replicate ArchR bug in handling nested genes}

\item{fragments}{Input fragments object}

\item{tile_max_count}{Maximum value in the tile counts matrix. If not null, tile counts higher than this will be clipped to \code{tile_max_count}.
Equivalent to \code{ceiling} argument of \code{ArchR::addGeneScoreMatrix()}}

\item{scale_factor}{If not null, counts for each cell will be scaled to sum to \code{scale_factor}. Equivalent to \code{scaleTo} argument of \code{ArchR::addGeneScoreMatrix()}}

\item{tile_matrix_path}{Path of a directory where the intermediate tile matrix will be saved}
}
\value{
\strong{gene_score_weights_archr}

Weight matrix of dimension genes x tiles

\strong{gene_score_archr}

Gene score matrix of dimension genes x cells.
}
\description{
Gene activity scores can be calculated as a distance-weighted sum of per-tile accessibility.
The tile weights for each gene can be represented as a sparse matrix of dimension genes x tiles.
If we multiply this weight matrix by a corresponding tile matrix (tiles x cells), then we can
get a gene activity score matrix of genes x cells. \code{gene_score_weights_archr()} calculates the
weight matrix (best if you have a pre-computed tile matrix), while \code{gene_score_archr()} provides
a easy-to-use wrapper.
}
\details{
\code{gene_score_weights_archr}
Given a set of tile coordinates and distances returned by \code{gene_score_tiles_archr()},
calculate a weight matrix of dimensions genes x tiles. This matrix can be
multiplied with a tile matrix to obtain ArchR-compatible gene activity scores.
}
