% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneScores.R
\name{geneScoreWeightsArchR}
\alias{geneScoreWeightsArchR}
\title{Calculate a sparse matrix of GeneActivityScore weights given a set of tile coordinates}
\usage{
geneScoreWeightsArchR(tiles, chr_sizes = GenomeInfoDb::seqlengths(tiles))
}
\arguments{
\item{tiles}{Tile coordinates as returned by geneScoreTilesArchR}

\item{chr_sizes}{Named vector of chromosome sizes with chromosmes in the order they are listed
in the tile matrix (used to calculate the index of each tile)}
}
\value{
Weight matrix of dimension genes x tiles
}
\description{
Calculate a sparse matrix of GeneActivityScore weights given a set of tile coordinates
}
