% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{import_matrix_market}
\alias{import_matrix_market}
\alias{import_matrix_market_10x}
\title{Import MatrixMarket files}
\usage{
import_matrix_market(
  mtx_path,
  outdir = tempfile("matrix_market"),
  row_names = NULL,
  col_names = NULL,
  row_major = FALSE,
  tmpdir = tempdir(),
  load_bytes = 4194304L,
  sort_bytes = 1073741824L
)

import_matrix_market_10x(
  mtx_dir,
  outdir = tempfile("matrix_market"),
  feature_type = NULL,
  row_major = FALSE,
  tmpdir = tempdir(),
  load_bytes = 4194304L,
  sort_bytes = 1073741824L
)
}
\arguments{
\item{mtx_path}{Path of mtx or mtx.gz file}

\item{outdir}{Directory to store the output}

\item{row_names}{Character vector of row names}

\item{col_names}{Character vector of col names}

\item{row_major}{If true, store the matrix in row-major orientation}

\item{tmpdir}{Temporary directory to use for intermediate storage}

\item{load_bytes}{The minimum contiguous load size during the merge sort passes}

\item{sort_bytes}{The amount of memory to allocate for re-sorting chunks of entries}

\item{mtx_dir}{Directory holding matrix.mtx.gz, barcodes.tsv.gz, and features.tsv.gz}

\item{feature_type}{String or vector of feature types to include. (cellranger 3.0 and newer)}
}
\value{
MatrixDir object with the imported matrix
}
\description{
Read a sparse matrix from a MatrixMarket file. This is a text-based format used by
10x, Parse, and others to store sparse matrices.
Format details on the \href{https://math.nist.gov/MatrixMarket/formats.html}{NIST website}.
}
\details{
Import MatrixMarket mtx files to the BPCells format. This implementation ensures
fixed memory usage even for very large inputs by doing on-disk sorts. It will be
much slower than hdf5 inputs, so only use MatrixMarket format when absolutely necessary.

As a rough speed estimate, importing the 17GB Parse
\href{https://www.parsebiosciences.com/datasets/pbmc/single-cell-rna-sequencing-of-1-million-human-cells-in-a-single-experiment}{1M PBMC}
\code{DGE_1M_PBMC.mtx} file takes about 4 minutes and 1.3GB of RAM, producing a compressed output matrix of 1.5GB. \code{mtx.gz}
files will be slower to import due to gzip decompression.

When importing from 10x mtx files, the row and column names can be read automatically
using the \code{import_matrix_market_10x()} convenience function.
}
