% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{IterableFragments-methods}
\alias{IterableFragments-methods}
\alias{show,IterableFragments-method}
\alias{cellNames}
\alias{cellNames<-}
\alias{chrNames}
\alias{chrNames<-}
\title{IterableFragments methods}
\usage{
\S4method{show}{IterableFragments}(object)

cellNames(x)

cellNames(x, ...) <- value

chrNames(x)

chrNames(x, ...) <- value
}
\arguments{
\item{object}{IterableFragments object}

\item{x}{an IterableFragments object}

\item{value}{Character vector of new names}
}
\value{
\itemize{
\item \code{cellNames()} Character vector of cell names, or NULL if none are known
}

\itemize{
\item \code{chrNames()}: Character vector of chromosome names, or NULL if none are known
}
}
\description{
Methods for IterableFragments objects
}
\details{
\itemize{
\item \verb{cellNames<-} It is only possible to replace names, not add new names.
}

\itemize{
\item \verb{chrNames<-} It is only possible to replace names, not add new names.
}
}
\section{Functions}{
\itemize{
\item \code{show(IterableFragments)}: Print IterableFragments

\item \code{cellNames()}: Get cell names

\item \code{cellNames(x, ...) <- value}: Set cell names

\item \code{chrNames()}: Set chromosome names

\item \code{chrNames(x, ...) <- value}: Set chromosome names

}}
\examples{
## Prep data
frags <- tibble::tibble(
  chr = paste0("chr", c(rep(1,3), rep(2,3))),
  start = seq(10, 260, 50),
  end = start + 30,
  cell_id = paste0("cell", c(rep(1, 2), rep(2,2), rep(3,2)))
) 
frags
frags <- frags \%>\% convert_to_fragments()


#######################################################################
## show(IterableFragments) example
#######################################################################
show(frags)


#######################################################################
## cellNames(IterableFragments) example
#######################################################################
cellNames(frags)


#######################################################################
## cellNames(IterableFragments)<- example
#######################################################################
cellNames(frags) <- paste0("cell", 5:7)
cellNames(frags)


#######################################################################
## chrNames(IterableFragments) example
#######################################################################
chrNames(frags)


#######################################################################
## chrNames(IterableFragments)<- example
#######################################################################
chrNames(frags) <- paste0("chr", 5:6)
chrNames(frags)


}
