% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R, R/matrix_stats.R, R/transforms.R
\name{IterableMatrix-methods-stats}
\alias{IterableMatrix-methods-stats}
\alias{rowSums,IterableMatrix-method}
\alias{colSums,IterableMatrix-method}
\alias{rowMeans,IterableMatrix-method}
\alias{colMeans,IterableMatrix-method}
\alias{colVars}
\alias{rowVars}
\alias{rowMaxs}
\alias{colMaxs}
\alias{rowQuantiles}
\alias{colQuantiles}
\alias{log1p,IterableMatrix-method}
\alias{log1p_slow}
\alias{expm1,IterableMatrix-method}
\alias{expm1_slow}
\title{IterableMatrix summary statistics}
\usage{
\S4method{rowSums}{IterableMatrix}(x)

\S4method{colSums}{IterableMatrix}(x)

\S4method{rowMeans}{IterableMatrix}(x)

\S4method{colMeans}{IterableMatrix}(x)

colVars(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  ...,
  useNames = TRUE
)

rowVars(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  ...,
  useNames = TRUE
)

rowMaxs(x, rows = NULL, cols = NULL, na.rm = FALSE, ..., useNames = TRUE)

colMaxs(x, rows = NULL, cols = NULL, na.rm = FALSE, ..., useNames = TRUE)

rowQuantiles(
  x,
  rows = NULL,
  cols = NULL,
  probs = seq(from = 0, to = 1, by = 0.25),
  na.rm = FALSE,
  type = 7L,
  digits = 7L,
  ...,
  useNames = TRUE,
  drop = TRUE
)

colQuantiles(
  x,
  rows = NULL,
  cols = NULL,
  probs = seq(from = 0, to = 1, by = 0.25),
  na.rm = FALSE,
  type = 7L,
  digits = 7L,
  ...,
  useNames = TRUE,
  drop = TRUE
)

\S4method{log1p}{IterableMatrix}(x)

log1p_slow(x)

\S4method{expm1}{IterableMatrix}(x)

expm1_slow(x)
}
\arguments{
\item{x}{IterableMatrix object or a matrix-like object.}

\item{rows}{(Integer) Optional vector of row indices to operate over.}

\item{cols}{(Integer) Optional vector of column indices to operate over.}

\item{na.rm}{(Logical) Should missing values (NA) be removed?}

\item{center}{Optional center values (vector of length nrow(x) or ncol(x))}

\item{...}{Additional arguments passed to methods}

\item{useNames}{(Logical) Whether to use row and column names in the output.}

\item{probs}{(Numeric) Quantile value(s) to be computed, between 0 and 1.}

\item{type}{(Integer) between 4 and 9 selecting which quantile algorithm to use, detailed in \code{matrixStats::rowQuantiles()}}

\item{digits}{Number of decimal places for quantile calculations}

\item{drop}{(Logical) If TRUE and only one quantile is requested, the result is coerced to a vector (For non-BPCells objects).}
}
\value{
\itemize{
\item \code{rowSums()}: vector of row sums
}

\itemize{
\item \code{colSums()}: vector of col sums
}

\itemize{
\item \code{rowMeans()}: vector of row means
}

\itemize{
\item \code{colMeans()}: vector of col means
}

\itemize{
\item \code{colVars()}: vector of col variance
}

\itemize{
\item \code{rowVars()}: vector of row variance
}

\itemize{
\item \code{rowMaxs()}: vector of maxes for every row
}

\itemize{
\item \code{colMaxs()}: vector of column maxes
}

\itemize{
\item \verb{rowQuantiles():} If \code{length(probs) == 1}, return a numeric with number of entries equal to the number of rows in the matrix.
Else, return a Matrix of quantile values, with cols representing each quantile, and each row representing a row in the input matrix.
}

\itemize{
\item \verb{colQuantiles():} If \code{length(probs) == 1}, return a numeric with number of entries equal to the number of columns in the matrix.
Else, return a Matrix of quantile values, with cols representing each quantile, and each row representing a col in the input matrix.
}
}
\description{
Summaries and reductions for IterableMatrix objects (sums, means, variances, quantiles, extrema).
}
\section{Functions}{
\itemize{
\item \code{rowSums(IterableMatrix)}: Calculate rowSums

\item \code{colSums(IterableMatrix)}: Calculate colSums

\item \code{rowMeans(IterableMatrix)}: Calculate rowMeans

\item \code{colMeans(IterableMatrix)}: Calculate colMeans

\item \code{colVars()}: Calculate colVars (replacement for \code{matrixStats::colVars()})

\item \code{rowVars()}: Calculate rowVars (replacement for \code{matrixStats::rowVars()}) (generic)

\item \code{rowMaxs()}: Calculate rowMaxs (replacement for \code{matrixStats::rowMaxs()}) (generic)

\item \code{colMaxs()}: Calculate colMax (replacement for \code{matrixStats::colMax()})

\item \code{rowQuantiles()}: Calculate rowQuantiles (replacement for \code{matrixStats::rowQuantiles})

\item \code{colQuantiles()}: Calculate colQuantiles (replacement for \code{matrixStats::colQuantiles})

\item \code{log1p(IterableMatrix)}: Calculate log(x + 1)

\item \code{log1p_slow()}: Calculate log(x + 1) (non-SIMD version)

\item \code{expm1(IterableMatrix)}: Calculate exp(x) - 1

\item \code{expm1_slow()}: Calculate exp(x) - 1 (non-SIMD version)

}}
\examples{
## Prep data
mat <- matrix(1:25, nrow = 5) \%>\% as("dgCMatrix")
mat
mat <- as(mat, "IterableMatrix")
mat

#######################################################################
## rowSums() example
#######################################################################
rowSums(mat)


#######################################################################
## colSums() example
#######################################################################
colSums(mat)


#######################################################################
## rowMeans() example
#######################################################################
rowMeans(mat)


#######################################################################
## colMeans() example
#######################################################################
colMeans(mat)


#######################################################################
## colVars() example
#######################################################################
colVars(mat)


#######################################################################
## rowVars() example
#######################################################################
rowVars(mat)


#######################################################################
## rowMaxs() example
#######################################################################
rowMaxs(mat)


#######################################################################
## colMaxs() example
#######################################################################
colMaxs(mat)


#######################################################################
## rowQuantiles() example
#######################################################################
rowQuantiles(transpose_storage_order(mat))


#######################################################################
## colQuantiles() example
#######################################################################
colQuantiles(mat)


#######################################################################
## log1p() example
#######################################################################
log1p(mat)


#######################################################################
## log1p_slow() example
#######################################################################
log1p_slow(mat)


#######################################################################
## expm1() example
#######################################################################
expm1(mat)


#######################################################################
## expm1_slow() example
#######################################################################
expm1_slow(mat)


}
