% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{IterableMatrix-methods}
\alias{IterableMatrix-methods}
\alias{show,IterableMatrix-method}
\alias{t,IterableMatrix-method}
\alias{\%*\%,IterableMatrix,matrix-method}
\alias{rowSums,IterableMatrix-method}
\alias{colSums,IterableMatrix-method}
\alias{rowMeans,IterableMatrix-method}
\alias{colMeans,IterableMatrix-method}
\title{IterableMatrix methods}
\usage{
\S4method{show}{IterableMatrix}(object)

\S4method{t}{IterableMatrix}(x)

\S4method{\%*\%}{IterableMatrix,matrix}(x, y)

\S4method{rowSums}{IterableMatrix}(x)

\S4method{colSums}{IterableMatrix}(x)

\S4method{rowMeans}{IterableMatrix}(x)

\S4method{colMeans}{IterableMatrix}(x)
}
\arguments{
\item{object}{IterableMatrix object}

\item{x}{IterableMatrix object}

\item{y}{matrix}
}
\value{
\itemize{
\item \code{t()} Transposed object
}

\itemize{
\item \code{x \%*\% y}: dense matrix result
}

\itemize{
\item \code{rowSums()}: vector of row sums
}

\itemize{
\item \code{colSums()}: vector of col sums
}

\itemize{
\item \code{rowMeans()}: vector of row means
}

\itemize{
\item \code{colMeans()}: vector of col means
}
}
\description{
Methods for IterableMatrix objects
}
\section{Functions}{
\itemize{
\item \code{show(IterableMatrix)}: Display an IterableMatrix

\item \code{t(IterableMatrix)}: Transpose an IterableMatrix

\item \code{x \%*\% y}: Multiply by a dense matrix

\item \code{rowSums(IterableMatrix)}: Calculate rowSums

\item \code{colSums(IterableMatrix)}: Calculate colSums

\item \code{rowMeans(IterableMatrix)}: Calculate rowMeans

\item \code{colMeans(IterableMatrix)}: Calculate colMeans

}}
