% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R, R/transforms.R
\name{IterableMatrix-methods}
\alias{IterableMatrix-methods}
\alias{matrix_type}
\alias{storage_order}
\alias{show,IterableMatrix-method}
\alias{t,IterableMatrix-method}
\alias{\%*\%,IterableMatrix,matrix-method}
\alias{rowSums,IterableMatrix-method}
\alias{colSums,IterableMatrix-method}
\alias{rowMeans,IterableMatrix-method}
\alias{colMeans,IterableMatrix-method}
\alias{log1p,IterableMatrix-method}
\alias{log1p_slow}
\alias{expm1,IterableMatrix-method}
\alias{expm1_slow}
\alias{^,IterableMatrix,numeric-method}
\alias{pow_slow}
\alias{<,numeric,IterableMatrix-method}
\alias{>,IterableMatrix,numeric-method}
\alias{<=,numeric,IterableMatrix-method}
\alias{>=,IterableMatrix,numeric-method}
\alias{round,IterableMatrix-method}
\alias{*,IterableMatrix,numeric-method}
\alias{+,IterableMatrix,numeric-method}
\alias{/,IterableMatrix,numeric-method}
\alias{-,IterableMatrix,numeric-method}
\title{IterableMatrix methods}
\usage{
matrix_type(x)

storage_order(x)

\S4method{show}{IterableMatrix}(object)

\S4method{t}{IterableMatrix}(x)

\S4method{\%*\%}{IterableMatrix,matrix}(x, y)

\S4method{rowSums}{IterableMatrix}(x)

\S4method{colSums}{IterableMatrix}(x)

\S4method{rowMeans}{IterableMatrix}(x)

\S4method{colMeans}{IterableMatrix}(x)

\S4method{log1p}{IterableMatrix}(x)

log1p_slow(x)

\S4method{expm1}{IterableMatrix}(x)

expm1_slow(x)

\S4method{^}{IterableMatrix,numeric}(e1, e2)

pow_slow(x, exponent)

\S4method{<}{numeric,IterableMatrix}(e1, e2)

\S4method{>}{IterableMatrix,numeric}(e1, e2)

\S4method{<=}{numeric,IterableMatrix}(e1, e2)

\S4method{>=}{IterableMatrix,numeric}(e1, e2)

\S4method{round}{IterableMatrix}(x, digits = 0)

\S4method{*}{IterableMatrix,numeric}(e1, e2)

\S4method{+}{IterableMatrix,numeric}(e1, e2)

\S4method{/}{IterableMatrix,numeric}(e1, e2)

\S4method{-}{IterableMatrix,numeric}(e1, e2)
}
\arguments{
\item{x}{IterableMatrix object}

\item{object}{IterableMatrix object}

\item{y}{matrix}
}
\value{
\itemize{
\item \code{t()} Transposed object
}

\itemize{
\item \code{x \%*\% y}: dense matrix result
}

\itemize{
\item \code{rowSums()}: vector of row sums
}

\itemize{
\item \code{colSums()}: vector of col sums
}

\itemize{
\item \code{rowMeans()}: vector of row means
}

\itemize{
\item \code{colMeans()}: vector of col means
}
}
\description{
Generic methods and built-in functions for IterableMatrix objects
}
\section{Functions}{
\itemize{
\item \code{matrix_type()}: Get the matrix data type (mat_uint32_t, mat_float, or mat_double for now)

\item \code{storage_order()}: Get the matrix storage order ("row" or "col")

\item \code{show(IterableMatrix)}: Display an IterableMatrix

\item \code{t(IterableMatrix)}: Transpose an IterableMatrix

\item \code{x \%*\% y}: Multiply by a dense matrix

\item \code{rowSums(IterableMatrix)}: Calculate rowSums

\item \code{colSums(IterableMatrix)}: Calculate colSums

\item \code{rowMeans(IterableMatrix)}: Calculate rowMeans

\item \code{colMeans(IterableMatrix)}: Calculate colMeans

\item \code{log1p(IterableMatrix)}: Calculate log(x + 1)

\item \code{log1p_slow()}: Calculate log(x + 1) (non-SIMD version)

\item \code{expm1(IterableMatrix)}: Calculate exp(x) - 1

\item \code{expm1_slow()}: Calculate exp(x) - 1 (non-SIMD version)

\item \code{e1^e2}: Calculate x^y (elementwise)

\item \code{pow_slow()}: Calculate x^y (elementwise, non-SIMD version)

\item \code{e1 < e2}: Binarize matrix according to numeric < matrix comparison

\item \code{e1 > e2}: Binarize matrix according to matrix > numeric comparison

\item \code{e1 <= e2}: Binarize matrix according to numeric <= matrix comparison

\item \code{e1 >= e2}: Binarize matrix according to matrix >= numeric comparison

\item \code{round(IterableMatrix)}: round to nearest integer (digits must be 0)

\item \code{e1 * e2}: Multiply by a constant, or multiply rows by a vector length nrow(mat)

\item \code{e1 + e2}: Add a constant, or row-wise addition with a vector length nrow(mat)

\item \code{e1 / e2}: Divide by a constant, or divide rows by a vector length nrow(mat)

\item \code{e1 - e2}: Subtract a constant, or row-wise subtraction with a vector length nrow(mat)

}}
