% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{knn_to_graph}
\alias{knn_to_graph}
\alias{knn_to_snn_graph}
\title{K Nearest Neighbor (KNN) Graph}
\usage{
knn_to_graph(knn, use_weights = FALSE, self_loops = TRUE)

knn_to_snn_graph(knn, min_val = 1/15, self_loops = TRUE)
}
\arguments{
\item{knn}{List of 2 matrices -- idx for cell x K neighbor indices,
dist for cell x K neighbor distances}

\item{use_weights}{boolean for whether to replace all distance weights with 1}

\item{self_loops}{boolean for whether to allow cells to count themselves as neighbors}

\item{min_val}{minimum jaccard index between neighbors. Values below this will
round to 0}
}
\value{
\strong{knn_to_graph}
Sparse matrix (dgCMatrix) where \code{mat[i,j]} = distance from cell \code{i} to
cell \code{j}, or 0 if cell \code{j} is not in the K nearest neighbors of \code{i}

\strong{knn_to_snn_graph}
Sparse matrix (dgCMatrix) where \code{mat[i,j]} = jaccard index of the overlap
in nearest neigbors between cell \code{i} and cell \code{j}, or 0 if the jaccard index
is < \code{min_val}
}
\description{
Convert a KNN object (e.g. returned by \code{knn_hnsw()} or \code{knn_annoy()}) into
a graph. The graph is represented as a sparse adjacency matrix.
}
\details{
\strong{knn_to_graph}
Create a knn graph

\strong{knn_to_snn_graph}
Convert a knn object into a shared nearest neighbors adjacency matrix.
This follows the algorithm that Seurat uses to compute SNN graphs
}
