% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\docType{class}
\name{LinearOperator-class}
\alias{LinearOperator-class}
\title{Represent a sparse matrix-vector product operation}
\description{
LinearOperators perform sparse matrix-vector product operations for
for downstream matrix solvers. They avoid repeatedly calling iterate_matrix
from an SVD solver for a possible efficiency gain
}
\keyword{internal}
