% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_project_predict.R
\docType{class}
\name{LSITransformer}
\alias{LSITransformer}
\title{Perform latent semantic indexing (LSI) on a matrix.}
\usage{
LSITransformer(z_score_norm, n_dimensions, scale_factor, threads)
}
\description{
Perform latent semantic indexing (LSI) on a matrix.

Create a new LSITransformer object
}
