% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{add_rows}
\alias{add_rows}
\alias{add_cols}
\alias{multiply_rows}
\alias{multiply_cols}
\title{Broadcasting vector arithmetic}
\usage{
add_rows(mat, vec)

add_cols(mat, vec)

multiply_rows(mat, vec)

multiply_cols(mat, vec)
}
\arguments{
\item{mat}{Matrix-like object}

\item{vec}{Numeric vector}
}
\value{
Matrix-like object
}
\description{
Convenience functions for adding or multiplying
each row / column of a matrix by a number.
}
