% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix_R_conversion_coercions}
\alias{matrix_R_conversion_coercions}
\alias{as.matrix.IterableMatrix}
\alias{as.matrix,IterableMatrix-method}
\alias{matrix_R_conversion}
\title{Convert between BPCells matrix and R objects.}
\usage{
\method{as.matrix}{IterableMatrix}(x, ...)
}
\arguments{
\item{x}{Matrix object to convert}

\item{...}{Additional arguments passed to methods}
}
\description{
BPCells matrices can be interconverted with Matrix package
dgCMatrix sparse matrices, as well as base R
dense matrices (though this may result in high memory usage for large matrices)
}
\details{
Use base R's \code{as()} to convert between BPCells matrices and
\code{dgCMatrix}/\code{matrix} representations, while \code{as.matrix()} materialises
dense matrices directly when needed.
}
\examples{
# setup data
mat <- matrix(1:25, nrow=5, ncol=5)
rownames(mat) <- paste0("gene", seq_len(5))
colnames(mat) <- paste0("cell", seq_len(5))
mat <- mat \%>\% as("IterableMatrix")
mat

#######################################################################
## as(bpcells_mat, "dgCMatrix") example
#######################################################################
mat_dgc <- as(mat, "dgCMatrix")
mat_dgc


## as.matrix(bpcells_mat) example
as.matrix(mat)

## Alternatively, can also use function as()
as(mat, "matrix")

#######################################################################
## as(dgc_mat, "IterableMatrix") example
#######################################################################
as(mat_dgc, "IterableMatrix")


}
