% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix_R_conversion}
\alias{matrix_R_conversion}
\title{Convert between BPCells matrix and R objects.}
\usage{
# Convert to R from BPCells
as(bpcells_mat, "dgCMatrix") # Sparse matrix conversion
as.matrix(bpcells_mat) # Dense matrix conversion

# Convert to BPCells from R
as(dgc_mat, "IterableMatrix")
}
\description{
BPCells matrices can be interconverted with Matrix package
dgCMatrix sparse matrices, as well as base R
dense matrices (though this may result in high memory usage for large matrices)
}
