% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{merge_cells}
\alias{merge_cells}
\title{Merge cells into pseudobulks}
\usage{
merge_cells(fragments, cell_groups)
}
\arguments{
\item{fragments}{Input fragments object}

\item{cell_groups}{Character or factor vector providing a group for each cell.
Ordering is the same as \code{cellNames(fragments)}}
}
\description{
Peak and tile matrix calculations can be sped up by reducing the number of
cells. For cases where the outputs are going to be added together afterwards, this can
provide a performance improvement
}
