% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{normalize_tfidf}
\alias{normalize_tfidf}
\title{Normalize a \verb{(features x cells)} matrix using term frequency-inverse document frequency.}
\usage{
normalize_tfidf(mat, feature_means = NULL, scale_factor = 10000, threads = 1L)
}
\arguments{
\item{mat}{(IterableMatrix) Matrix to normalize.}

\item{feature_means}{(numeric) Means of the features to normalize by.  If no names are provided, then
each numeric value is assumed to correspond to the feature mean for the corresponding row of the matrix.
Else, map each feature name to its mean value.}

\item{scale_factor}{(numeric) Scale factor to multiply matrix by for log normalization.}

\item{threads}{(integer) Number of threads to use.s}
}
\value{
tf-idf normalized matrix.  For each element \eqn{x_{ij}} in matrix \eqn{X} with \eqn{i} features and \eqn{j} cells,
the tf-idf normalization of that element, \eqn{\tilde{x}_{ij}} is calculated as:
\eqn{\tilde{x}_{ij} = \log(\frac{x_{ij} \cdot \text{scaleFactor}}{\text{rowMean}_i\cdot \text{colSum}_j} + 1)}
}
\description{
Normalize a \verb{(features x cells)} matrix using term frequency-inverse document frequency.
}
