% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{open_fragments_10x}
\alias{open_fragments_10x}
\alias{write_fragments_10x}
\title{Read/write a 10x fragments file}
\usage{
open_fragments_10x(path, comment = "#", end_inclusive = TRUE)

write_fragments_10x(
  fragments,
  path,
  end_inclusive = TRUE,
  append_5th_column = FALSE
)
}
\arguments{
\item{path}{File path (e.g. fragments.tsv or fragments.tsv.gz)}

\item{comment}{Skip lines at beginning of file which start with comment string}

\item{end_inclusive}{Whether the end coordinate of the bed is inclusive -- i.e. there was an
insertion at the end coordinate rather than the base before the end coordinate. This is the
10x default, though it's not quite standard for the bed file format.}

\item{fragments}{Input fragments object}

\item{append_5th_column}{Whether to include 5th column of all 0 for compatibility
with 10x fragment file outputs (defaults to 4 columns chr,start,end,cell)}
}
\value{
10x fragments file object
}
\description{
10x fragment files come in a bed-like format, with columns chr, start, end,
cell_id, and pcr_duplicates. Unlike a standard bed format, the format from
cellranger has an \emph{inclusive} end-coordinate, meaning the end coordinate itself
is what should be counted as the tagmentation site, rather than offset by 1.
}
\details{
\strong{open_fragments_10x}

No disk operations will take place until the fragments are used in a function

\strong{write_fragments_10x}

Fragments will be written to disk immediately, then returned in a readable object.
}
\examples{
## Download example fragments from pbmc 500 dataset and save in temp directory
data_dir <- file.path(tempdir(), "frags_10x")
dir.create(data_dir, recursive = TRUE, showWarnings = FALSE)
url_base <- "https://cf.10xgenomics.com/samples/cell-atac/2.0.0/atac_pbmc_500_nextgem/"
frags_file <- "atac_pbmc_500_nextgem_fragments.tsv.gz"
atac_raw_url <- paste0(url_base, frags_file)
if (!file.exists(file.path(data_dir, frags_file))) {
 download.file(atac_raw_url, file.path(data_dir, frags_file), mode="wb")
}

#######################################################################
## open_fragments_10x() example
#######################################################################
frags <- open_fragments_10x(
 file.path(data_dir, frags_file)
)
## A Fragments object imported from 10x will not have cell/chromosome 
## information directly known unless written as a BPCells fragment object
frags

frags \%>\% write_fragments_dir(
 file.path(data_dir, "demo_frags_from_h5"), 
 overwrite = TRUE
)


#######################################################################
## write_fragments_10x() example
#######################################################################
frags <- write_fragments_10x(
 frags,
 file.path(data_dir, paste0("new_", frags_file))
)
frags


}
