% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeAnnotations.R
\name{order_ranges}
\alias{order_ranges}
\title{Get end-sorted ordering for genome ranges}
\usage{
order_ranges(ranges, chr_levels, sort_by_end = TRUE)
}
\arguments{
\item{ranges}{List, GRanges, or data.frame object. Must have start, end, and chr defined. See \link{genomic-ranges}}

\item{chr_levels}{Ordering of chromosome names}

\item{sort_by_end}{If TRUE (defualt), sort by (chr, end, start). Else sort by (chr, start, end)}
}
\value{
Numeric vector analagous to the \code{order} function. Provides an index
selection that will reorder the input ranges to be sorted by chr, end, start
}
\description{
Use this function to order regioins prior to calling \code{peak_matrix()} or \code{tile_matrix()}.
}
