% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeAnnotations.R
\name{order_ranges}
\alias{order_ranges}
\title{Get end-sorted ordering for genome ranges}
\usage{
order_ranges(ranges, chr_levels, sort_by_end = TRUE)
}
\arguments{
\item{ranges}{Genomic regions given as GRanges, data.frame, or list. See \code{help("genomic-ranges-like")} for details on format and coordinate systems. Required attributes:
\itemize{
\item \code{chr}, \code{start}, \code{end}: genomic position
}}

\item{chr_levels}{Ordering of chromosome names}

\item{sort_by_end}{If TRUE (defualt), sort by (chr, end, start). Else sort by (chr, start, end)}
}
\value{
Numeric vector analagous to the \code{order} function. Provides an index
selection that will reorder the input ranges to be sorted by chr, end, start
}
\description{
Use this function to order regioins prior to calling \code{peak_matrix()} or \code{tile_matrix()}.
}
