% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{peak_matrix}
\alias{peak_matrix}
\title{Calculate ranges x cells overlap matrix}
\usage{
peak_matrix(
  fragments,
  ranges,
  zero_based_coords = !is(ranges, "GRanges"),
  explicit_peak_names = TRUE
)
}
\arguments{
\item{fragments}{Input fragments object. Must have cell names and chromosome names defined}

\item{ranges}{GRanges object with the ranges to overlap, or list/data frame with columns chr, start, & end.
Must be sorted in order (chr, end, start) where chromosomes are ordered according to the chromosome names of \code{fragments}.}

\item{zero_based_coords}{Whether to convert the ranges from a 1-based end-inclusive
coordinate system to a 0-based end-exclusive coordinate system. Defaults to true
for GRanges and false for other formats
(see \url{http://genome.ucsc.edu/blog/the-ucsc-genome-browser-coordinate-counting-systems/})}

\item{explicit_peak_names}{Boolean for whether to add rownames to the output matrix in format e.g
chr1:500-1000, where start and end coords are given in a 0-based coordinate system.
Note that either way, peak names will be written when the matrix is saved.}
}
\value{
Iterable matrix object with dimension ranges x cells. When saved,
the column names of the output matrix will be in the format chr1:500-1000,
where start and end coords are given in a 0-based coordinate system.
}
\description{
Calculate ranges x cells overlap matrix
}
\note{
When calculating the matrix directly from a fragments tsv, it's necessary to first call \code{select_chromosomes} in order to
provide the ordering of chromosomes to expect while reading the tsv.
}
