% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_project_predict.R
\docType{class}
\name{Pipeline}
\alias{Pipeline}
\title{S4 class for combining multiple pipeline steps into a single pipeline}
\usage{
Pipeline(steps = list())
}
\arguments{
\item{steps}{A list of ordered steps to be executed in the pipeline.}
}
\value{
A new Pipeline object.
}
\description{
S4 class for combining multiple pipeline steps into a single pipeline

Return a new Pipeline object
}
\section{Slots}{

\describe{
\item{\code{steps}}{(list) List of pipeline steps to execute in order}
}}

