% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_project_predict.R
\docType{class}
\name{PipelineStep}
\alias{PipelineStep}
\title{S4 Class Representing a single transformer or predictor}
\description{
S4 Class Representing a single transformer or predictor
}
\section{Slots}{

\describe{
\item{\code{step_name}}{(character) Name of the step}

\item{\code{fitted}}{(logical) Whether the pipeline has been fitted}
}}

