% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_dot}
\alias{plot_dot}
\title{Dotplot}
\usage{
plot_dot(
  source,
  features,
  groups,
  group_order = NULL,
  gene_mapping = human_gene_mapping,
  colors = c("lightgrey", "#4682B4"),
  return_data = FALSE,
  apply_styling = TRUE
)
}
\arguments{
\item{source}{Feature x cell matrix or data.frame with features. For best results,
features should be sparse and log-normalized (e.g. run \code{log1p()} so zero raw counts
map to zero)}

\item{features}{Character vector of features to plot}

\item{groups}{Vector with one entry per cell, specifying the cell's group}

\item{group_order}{Optional vector listing ordering of groups}

\item{gene_mapping}{An optional vector for gene name matching with match_gene_symbol().}

\item{colors}{Color scale for plot}

\item{return_data}{If true, return data from just before plotting rather than a plot.}

\item{apply_styling}{If false, return a plot without pretty styling applied}
}
\description{
Plot feature levels per group or cluster as a grid of dots.
Dots are colored by z-score normalized average expression, and sized by percent non-zero.
}
