% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_tss_profile}
\alias{plot_tss_profile}
\title{Plot TSS profile}
\usage{
plot_tss_profile(
  fragments,
  genes,
  cell_groups = rlang::rep_along(cellNames(fragments), "all"),
  flank = 2000L,
  smooth = 0L,
  zero_based_coords = !is(genes, "GRanges"),
  colors = discrete_palette("stallion"),
  return_data = FALSE,
  apply_styling = TRUE
)
}
\arguments{
\item{fragments}{IterableFragments object}

\item{genes}{Coordinate ranges for genes (must include strand)}

\item{cell_groups}{Character or factor assigning a group to each cell, in order of
\code{cellNames(fragments)}}

\item{flank}{Number of flanking basepairs to include on either side of the motif}

\item{smooth}{Number of bases to smooth over (rolling average)}

\item{zero_based_coords}{If true, coordinates start and 0 and the end coordinate is not included in the range.
If false, coordinates start at 1 and the end coordinate is included in the range}

\item{return_data}{If true, return data from just before plotting rather than a plot.}

\item{apply_styling}{If false, return a plot without pretty styling applied}
}
\description{
Plot the enrichmment of insertions relative to transcription start sites (TSS).
Typically, this plot shows strong enrichment of insertions near a TSS, and a
small bump downstream around 220bp downstream of the TSS for the +1 nucleosome.
}
\seealso{
\code{footprint()}, \code{plot_tf_footprint()}
}
