% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{prefix_cell_names}
\alias{prefix_cell_names}
\title{Add sample prefix to cell names}
\usage{
prefix_cell_names(fragments, prefix)
}
\arguments{
\item{fragments}{Input fragments object.}

\item{prefix}{String to add as the prefix}
}
\value{
Fragments object with prefixed names
}
\description{
Rename cells by adding a prefix to the names. Most commonly this will be a sample name.
All cells will recieve the exact text of \code{prefix} added to the beginning, so any separator
characters like "_" must be included in the given \code{prefix}. Use this prior to merging
fragments from different experiments with \code{c()} in order to help prevent cell name
clashes.
}
