% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{prepare_demo_data}
\alias{prepare_demo_data}
\title{Create a small demo matrix and fragment object}
\usage{
prepare_demo_data(
  directory = NULL,
  mat_name = NULL,
  frags_name = NULL,
  timeout = 300,
  remove_input_data = TRUE
)
}
\arguments{
\item{directory}{(character) Where the input/output data should be stored.  If NULL, a temporary directory is created.}

\item{mat_name}{(character) Name of the RNA matrix file. If NULL, the matrix is named "demo_mat."}

\item{frags_name}{(character) Name of the ATAC fragments file. If NULL, the fragments are named "demo_frags".}

\item{timeout}{(numeric) Timeout for downloading files in seconds.}

\item{remove_input_data}{(logical) Whether to remove the downloaded non-procesed matrix, frags, gencode transcripts, and gencode genes
after processing.}
}
\value{
(list) A list with the RNA matrix under the name "mat", and the ATAC fragments under the name "frags".
}
\description{
Downloads a 10x Genomics dataset, then performs QC and subsetting.  Holds subsetted objects in disk,
and returns a list with both the matrix and fragments.
}
\details{
This function downloads the 10x Genomics PBMC 3k dataset, and filters the fragments and matrix to cells with at least 1000 reads.
Following, both fragments and the matrix is subset to only genes and insertions on chromosomes 4 and 11.
The RNA matrix is 1 MB and the fragments are 12.5 MB, after BPCells compression.
}
\keyword{internal}
