% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{prepare_demo_data}
\alias{prepare_demo_data}
\title{Create a small demo matrix and fragment object.}
\usage{
prepare_demo_data(
  directory = NULL,
  filter_qc = TRUE,
  subset = TRUE,
  timeout = 300
)
}
\arguments{
\item{directory}{(character) The directory where all the input/output data will be stored.
Downloaded intermediates will be stored in subdir \code{intermediates}.
If \code{NULL}, a temporary directory is created.}

\item{filter_qc}{(bool) Whether to filter both the RNA and ATAC data using QC information.}

\item{subset}{(bool) Whether to subset to only genes/insertions on chromosome 4 and 11.}

\item{timeout}{(numeric) Timeout for downloading files in seconds.}
}
\value{
(list) A list with the RNA matrix under the name \code{mat}, and the ATAC fragments under the name \code{frags}.
}
\description{
Downloads a
\href{https://www.10xgenomics.com/datasets/pbmc-from-a-healthy-donor-granulocytes-removed-through-cell-sorting-3-k-1-standard-2-0-0}{10x Genomics dataset},
consisting of 3k cells then performs optional QC and subsetting.  Holds subsetted objects in disk,
and returns a list with both the matrix and fragments.
}
\details{
This function downloads the 10x Genomics PBMC 3k dataset.
Filtering using QC information on the fragments and matrix provides cells with at least 1000 reads, 1000 frags, and a minimum tss enrichment of 10.
Subsetting provides only genes and insertions on chromosomes 4 and 11.
The name of the matrix and fragments folders are \code{demo_mat} and \code{demo_frags} respectively.
Additionally, choosing to qc filter appends a \verb{_filtered}, and choosing to subset data appends a \verb{_subsetted} to the name.
}
\keyword{internal}
