% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell_utils.R
\name{pseudobulk_matrix}
\alias{pseudobulk_matrix}
\title{Aggregate counts matrices by cell group or feature.}
\usage{
pseudobulk_matrix(
  mat,
  cell_groups,
  method = c("sum", "mean", "var", "non-zeros"),
  clip_values = FALSE,
  threads = 1L
)
}
\arguments{
\item{mat}{IterableMatrix object of dimensions features x cells}

\item{method}{(string) Method to aggregate counts.  Current options are:
\itemize{
\item \code{sum}
\item \code{mean}
}}

\item{clip_values}{(logical) If TRUE, clip high values to the 99th percentile of each cell.}

\item{threads}{(integer) Number of threads to use.}
}
\value{
(dgCMatrix) A Matrix row corresponding to features, and columns
coresponding to the aggregated counts, with each column named by the cell group.
}
\description{
Given a features x cells matrix, group cells by \code{cell_groups} and aggregate counts by \code{method} for each
feature.
}
