% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell_utils.R
\name{pseudobulk_matrix}
\alias{pseudobulk_matrix}
\title{Aggregate counts matrices by cell group or feature.}
\usage{
pseudobulk_matrix(
  mat,
  cell_groups,
  method = c("sum", "mean", "var", "non-zeros"),
  clip_values = FALSE,
  threads = 1L
)
}
\arguments{
\item{mat}{IterableMatrix object of dimensions features x cells}

\item{method}{(string) Method to aggregate counts.  Current options are:
\itemize{
\item \code{non-zeros}
\item \code{sum}
\item \code{mean}
\item \code{var}
}}

\item{clip_values}{(logical) If TRUE, clip high values to the 99th percentile of each cell.}

\item{threads}{(integer) Number of threads to use.}
}
\value{
(Named list) A list of matrices with each matrix representing a pseudobulk matrix with a different aggregation method.
Each matrix is of shape (features x groups), and names are one of \code{non_zeros}, \code{sum}, \code{mean}, \code{var}.
}
\description{
Given a features x cells matrix, group cells by \code{cell_groups} and aggregate counts by \code{method} for each
feature.
}
\details{
The stats are ordered by complexity: nonzero, sum, mean, then variance. All
less complex stats are calculated in the process of calculating a more complicated stat.
So to calculate mean and variance simultaneously, just ask for variance,
which will compute mean and nonzero counts as a side-effect.
}
