% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeAnnotations.R
\name{read_ucsc_chrom_sizes}
\alias{read_ucsc_chrom_sizes}
\title{Read UCSC chromosome sizes}
\usage{
read_ucsc_chrom_sizes(
  dir,
  genome = c("hg38", "mm39", "mm10", "mm9", "hg19"),
  keep_chromosomes = "chr[0-9]+|chrX|chrY",
  timeout = 300
)
}
\description{
Read chromosome sizes from UCSC and return as a tibble with one row per
chromosome.
The underlying data is pulled from here: \url{https://hgdownload.soe.ucsc.edu/downloads.html}
}
