% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{regress_out}
\alias{regress_out}
\title{Linear Regression}
\usage{
regress_out(mat, latent_data = NULL)
}
\arguments{
\item{mat}{An IterableMatrix}

\item{latent_data}{Extra data to regress out, should be a \code{data.frame} whose row number is equal to the column number of \code{mat}.}
}
\value{
An IterableMatrix. If input an empty \code{latent_data}, will return the \code{mat} itself.
}
\description{
Calculate the residuals from a linear model fit.
}
