% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rotate_x_labels}
\alias{rotate_x_labels}
\title{Rotate ggplot x axis labels}
\usage{
rotate_x_labels(degrees = 45)
}
\arguments{
\item{degrees}{Number of degrees to rotate by}
}
\description{
Rotate ggplot x axis labels
}
