% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{select_cells}
\alias{select_cells}
\title{Subset, translate, or reorder cell IDs}
\usage{
select_cells(fragments, cell_selection)
}
\arguments{
\item{fragments}{Input fragments object}

\item{cell_selection}{List of chromosme IDs (numeric), or names (character).
The output cell ID \code{n} will be taken
from the input cell with ID/name \code{cell_selection[n]}.}
}
\description{
Subset, translate, or reorder cell IDs
}
