% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{select_chromosomes}
\alias{select_chromosomes}
\title{Select chromosmes for subsetting or translating chromosome IDs in a fragments object}
\usage{
select_chromosomes(fragments, chromosome_selection)
}
\arguments{
\item{fragments}{Input fragments object}

\item{chromosome_selection}{List of chromosme IDs (numeric), or names (character).
The output chromosome ID n will be taken
from the input fragments chromosome with ID/name chromosome_selection\link{n}.}
}
\description{
Select chromosmes for subsetting or translating chromosome IDs in a fragments object
}
