% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{shift_fragments}
\alias{shift_fragments}
\title{Shift start or end coordinates}
\usage{
shift_fragments(fragments, shift_start = 0L, shift_end = 0L)
}
\arguments{
\item{fragments}{Input fragments object}

\item{shift_start}{How many basepairs to shift the start coords}

\item{shift_end}{How many basepairs to shift the end coords}
}
\value{
Shifted fragments object
}
\description{
Shifts start or end of fragments by a fixed amount, which can be useful to
correct the Tn5 offset.
}
\details{
The correct Tn5 offset is +/- 4bp since the Tn5 cut sites on opposite strands
are offset by 9bp. However, +4/-5 bp is often applied to bed-format files,
since the end coordinate in bed files is 1 past the last basepair of the sequenced
DNA fragment. This results in a bed-like format except with inclusive end coordinates.
}
\examples{
## Prep data
frags <- tibble::tibble(
  chr = "chr1",
  start = seq(10, 260, 50),
  end = start + 30,
  cell_id = paste0("cell1")
) \%>\% as("GRanges")
frags

frags <- frags \%>\% convert_to_fragments()


## Shift fragments
shift_fragments(frags, shift_start = 4, shift_end = -4) \%>\% as("GRanges")
}
