% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{subset_lengths}
\alias{subset_lengths}
\title{Subset fragments by length}
\usage{
subset_lengths(fragments, min_len = 0L, max_len = NA_integer_)
}
\arguments{
\item{fragments}{Input fragments object}

\item{min_len}{Minimum bases in fragment (inclusive)}

\item{max_len}{Maximum bases in fragment (inclusive)}
}
\value{
Fragments object
}
\description{
Subset fragments by length
}
\details{
Fragment length is calculated as \code{end-start}
}
