% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplots.R
\name{trackplot_combine}
\alias{trackplot_combine}
\title{Combine track plots}
\usage{
trackplot_combine(
  tracks,
  side_plot = NULL,
  title = NULL,
  side_plot_width = 0.3
)
}
\arguments{
\item{tracks}{List of tracks in order from top to bottom, generally ggplots as output from
the other \verb{trackplot_*()} functions.}

\item{side_plot}{Optional plot to align to the right (e.g. RNA expression per cluster). Will be aligned to a
\code{trackplot_coverage()} output if present, or else the first generic ggplot in the alignment. Should be in horizontal orientation and
in the same cluster ordering as the coverage plots.}

\item{title}{Text for overarching title of the plot}

\item{side_plot_width}{Fraction of width that should be used for the side plot relative to the main track area}
}
\value{
A plot object with aligned genome plots. Each aligned row has
the text label, y-axis, and plot body. The relative height of each row is given
by heights. A shared title and x-axis are put at the top.
}
\description{
Combines multiple track plots of the same region into a single grid.
Uses the \code{patchwork} package to perform the alignment.
}
\seealso{
\code{trackplot_coverage()}, \code{trackplot_gene()}, \code{trackplot_loop()}, \code{trackplot_scalebar()}
}
