% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplots.R
\name{trackplot_create_arrow_segs}
\alias{trackplot_create_arrow_segs}
\title{Break up segments into smaller segments the length of the plot, divided by size}
\usage{
trackplot_create_arrow_segs(data, region, size = 50, head_only = FALSE)
}
\arguments{
\item{data}{Dataframe of full segments to be broken up}

\item{region}{Region to be plotted with end and start attr}

\item{size}{int Number of arrows to span the x axis of track}

\item{head_only}{bool If TRUE, only the head of the segment will be plotted}
}
\value{
Dataframe of segments broken up into smaller segments.  Has columns start, end, and any additional metadata columns in original data
}
\description{
Break up segments into smaller segments the length of the plot, divided by size
}
\keyword{internal}
