% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplots.R
\name{trackplot_gene}
\alias{trackplot_gene}
\title{Plot transcript models}
\usage{
trackplot_gene(
  transcripts,
  region,
  exon_size = 2.5,
  gene_size = 0.5,
  label_size = 11 * 0.8/ggplot2::.pt,
  track_label = "Genes",
  return_data = FALSE
)
}
\arguments{
\item{transcripts}{Transcipt features given as GRanges, data.frame, or list. See \code{help("genomic-ranges-like")} for details on format and coordinate systems. Required attributes:
\itemize{
\item \code{chr}, \code{start}, \code{end}: genomic position
\item \code{strand}: +/- or TRUE/FALSE for positive or negative strand
\item \code{feature}: Only entries marked as \code{"transcript"} or \code{"exon"} will be considered
\item \code{gene_name}: Symbol or gene ID to display
\item \code{transcript_id}: Transcritp identifier to link transcripts and exons
}

Usually given as the output from \code{read_gencode_transcripts()}}

\item{region}{Region to plot, e.g. output from \code{gene_region()}. String of format "chr1:100-200",
or list/data.frame/GRanges of length 1 specifying chr, start, end. See \code{help("genomic-ranges-like")} for details}

\item{exon_size}{size for exon lines in units of mm}

\item{gene_size}{size for intron/gene lines in units of mm}

\item{label_size}{size for transcript labels in units of mm}

\item{return_data}{If true, return data from just before plotting rather than a plot.}

\item{labels}{Character vector with labels for each item in transcripts. NA for items that should not be labeled}

\item{transcript_size}{size for transcript lines in units of mm}
}
\value{
Plot of gene locations
}
\description{
Plot transcript models
}
\examples{
## Prep data
transcripts <- read_gencode_transcripts(
  file.path(tempdir(), "references"), release = "42",
  annotation_set = "basic", features = "transcript"
)
region <- "chr4:3034877-4034877"


## Plot gene trackplot
plot <- trackplot_gene(transcripts, region)
BPCells:::render_plot_from_storage(plot, width = 6, height = 1)
}
\seealso{
\code{trackplot_combine()}, \code{trackplot_coverage()}, \code{trackplot_loop()}, \code{trackplot_scalebar()}
}
