% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplots.R
\name{trackplot_gene}
\alias{trackplot_gene}
\title{Plot transcript models}
\usage{
trackplot_gene(
  transcripts,
  region,
  exon_size = 3,
  gene_size = 1,
  label_size = 3,
  return_data = FALSE
)
}
\arguments{
\item{transcripts}{GRanges, list, or data.frame of transcript features to plot.
Required attributes are:
\itemize{
\item \code{chr}, \code{start}, \code{end}: genomic position
\item \code{strand}: "+"/"-" or TRUE/FALSE for positive or negative strand
\item \code{feature} (only entries marked as \code{"transcript"} or \code{"exon"} will be considered)
\item \code{transcript_id}
\item \code{gene_name}
See \link{genomic-ranges} for more details
}}

\item{region}{GRanges of length 1 with region to plot, or list/data.frame with
one entry each for chr, start, end. See \code{gene_region()} or \link{genomic-ranges} for details}

\item{exon_size}{size for exon lines}

\item{label_size}{size for transcript labels}

\item{return_data}{If true, return data from just before plotting rather than a plot.}

\item{labels}{Character vector with labels for each item in transcripts. NA for items that should not be labeled}

\item{transcript_size}{size for transcript lines}
}
\value{
Plot of gene locations
}
\description{
Plot transcript models
}
