% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackplots.R
\name{set_trackplot_label}
\alias{set_trackplot_label}
\alias{set_trackplot_height}
\alias{get_trackplot_height}
\title{Adjust trackplot properties}
\usage{
set_trackplot_label(plot, labels)

set_trackplot_height(plot, height)

get_trackplot_height(plot)
}
\arguments{
\item{plot}{ggplot object}

\item{labels}{character vector of labels -- must match existing number of facets in plot}

\item{height}{New height. If numeric, adjusts relative height. If \code{ggplot2::unit} or \code{grid::unit} sets absolute height in specified units.
\code{"null"} units are interpreted as relative height.}
}
\value{
\strong{set_trackplot_label}: ggplot object with adjusted facet labels

\strong{set_trackplot_height}: ggplot object with adjusted trackplot height

\strong{get_trackplot_height}: \code{ggplot2::unit} object with height setting
}
\description{
Adjust labels and heights on trackplots. Labels are set as facet labels in ggplot2, and heights
are additional properties read by \code{trackplot_combine()} to determine relative height of input plots.
}
