% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{wrapMatrix}
\alias{wrapMatrix}
\title{Construct an S4 matrix object wrapping another matrix object}
\usage{
wrapMatrix(class, m, ...)
}
\description{
Helps to avoid duplicate storage of dimnames
}
\keyword{internal}
