% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac_utils.R
\name{write_insertion_bedgraph}
\alias{write_insertion_bedgraph}
\title{Write insertion counts to bedgraph file}
\usage{
write_insertion_bedgraph(
  fragments,
  path,
  cell_groups = rlang::rep_along(cellNames(fragments), "all"),
  insertion_mode = c("both", "start_only", "end_only")
)
}
\arguments{
\item{fragments}{IterableFragments object}

\item{path}{(character vector) Path(s) to save bedgraph to, optionally ending in ".gz" to add gzip compression. If \code{cell_groups} is provided,
\code{path} must be a named character vector, with one name for each level in \code{cell_groups}}

\item{cell_groups}{Character or factor assigning a group to each cell, in order of
\code{cellNames(fragments)}}

\item{insertion_mode}{(string) Which fragment ends to use for coverage calculation. One of "both", "start_only", or "end_only"}
}
\description{
Write insertion counts data for one or more pseudobulks to bedgraph format. This reports the total
number insertions at each basepair for each group listed in \code{cell_groups}.
}
\examples{
## Prep data
frags <- get_demo_frags()
bedgraph_outputs <- file.path(tempdir(), "bedgraph_outputs", "all.tar.gz")


## Write insertions
write_insertion_bedgraph(frags, file.path(bedgraph_outputs, "all.tar.gz"))
list.files(bedgraph_outputs)
}
